chore - split Body IR lowering into per-construct modules (#1101)
`body_ir.rs` was one 6.3k-line module whose `BodyBuilder` impl held every
construct's lowering. The remaining #1101 children each target a different
construct, so they all landed in that one impl and conflicted with each other by
construction.

The impl is now split along construct boundaries, each in its own file: `stmt`,
`control_flow`, `comprehensions`, `operators`, `expr`, `calls`, `literals`,
`match_`, `closures`, `async_`, `assertions`, and `defaults`. `body_ir.rs` keeps
the builder itself, its scaffolding (scopes, locals, temps, ownership, drops),
the module-level types, and the free helpers. It drops from 6,291 to ~2,550
lines.

Every method is a pure move. Cross-module calls need visibility, so methods in
the new modules carry `pub(super)`, which reaches only within `frontend::body_ir`
-- `BodyBuilder` is itself private, so nothing new is reachable outside.

The compatibility registry caught this move: eleven Body-IR anchors pointed at
`body_ir.rs` and went dangling. Rather than repointing each feature at a
specific new file, an anchor now resolves across its module's own tree. A
recorded path names a *module*, and `foo.rs` plus `foo/` are one module in Rust,
so a selector that moved into a submodule has not left the surface the anchor
records. The search does not recurse past that tree, so a selector that genuinely
leaves the surface still reports as dangling.
