+++ src/frontend/body_ir.rs
@@ -1142,3765 +1142,6 @@ impl<'type_info, 'source> BodyBuilder<'type_info, 'source> {
     }
 
     // ---- Statements ----
-
-    /// Lower every statement in `stmts` into `out`, within `scope`. Statements are lowered in source order and each
-    /// one is given the statement suffix that follows it (`&stmts[index + 1..]`), so last-use countdowns seeded by
-    /// [`Self::declare_new_local`] only count reads that can still occur after the declaration.
-    fn lower_block_into(
-        &mut self,
-        stmts: &[ast::Spanned<ast::Statement>],
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) {
-        for (index, stmt) in stmts.iter().enumerate() {
-            self.lower_stmt_into(stmt, &stmts[index + 1..], scope, out);
-        }
-    }
-
-    /// Lower one statement into `out`, dispatching on its AST kind. `remaining` is the statement suffix following
-    /// `stmt` in its enclosing block, threaded through to [`Self::lower_assignment`] for last-use seeding. Statement
-    /// kinds outside v0's covered subset fall through to an explicit [`Self::push_unsupported_stmt`] rather than
-    /// panicking (see this module's module-level docs for the exact covered/uncovered split).
-    fn lower_stmt_into(
-        &mut self,
-        stmt: &ast::Spanned<ast::Statement>,
-        remaining: &[ast::Spanned<ast::Statement>],
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) {
-        let span = hir_span(stmt.span);
-        match &stmt.node {
-            ast::Statement::Assignment(assignment) => self.lower_assignment(assignment, remaining, scope, span, out),
-            ast::Statement::FieldAssignment(field_assignment) => {
-                self.lower_field_assignment(field_assignment, scope, span, out)
-            }
-            ast::Statement::IndexAssignment(index_assignment) => {
-                self.lower_index_assignment(index_assignment, scope, span, out)
-            }
-            ast::Statement::CompoundAssignment(compound_assignment) => {
-                self.lower_compound_assignment(compound_assignment, scope, span, out)
-            }
-            ast::Statement::TupleUnpack(tuple_unpack) => {
-                self.lower_tuple_unpack(tuple_unpack, remaining, scope, span, out)
-            }
-            ast::Statement::TupleAssign(tuple_assign) => self.lower_tuple_assign(tuple_assign, scope, span, out),
-            ast::Statement::ChainedAssignment(chained_assignment) => {
-                self.lower_chained_assignment(chained_assignment, remaining, scope, span, out)
-            }
-            ast::Statement::Return(value) => {
-                let value = value.as_ref().map(|v| self.lower_expr_to_operand(v, scope, out));
-                out.push(bir::Statement {
-                    kind: bir::StatementKind::Return { value },
-                    span,
-                });
-            }
-            ast::Statement::If(if_stmt) => self.lower_if(if_stmt, scope, span, out),
-            ast::Statement::While(while_stmt) => self.lower_while(while_stmt, scope, span, out),
-            ast::Statement::For(for_stmt) => self.lower_for(for_stmt, scope, span, out),
-            ast::Statement::Expr(expr) => {
-                // `yield value` parses as an ordinary expression statement wrapping `ast::Expr::Yield(Some(_))`
-                // (there is no separate `ast::Statement::Yield` AST node) -- mirror the existing Rust-emission
-                // backend's own `lower_statement` (`src/backend/ir/lower/stmt.rs`), which special-cases this exact
-                // shape before falling back to generic expression-statement lowering. A bare `yield` (no value)
-                // falls through to the generic `Expr` arm below, same as that backend, and lowers via the
-                // expression-position `yield` stub (see the module docs).
-                if let ast::Expr::Yield(Some(value)) = &expr.node {
-                    self.lower_yield(value, scope, span, out);
-                } else {
-                    let value = self.lower_expr_to_operand(expr, scope, out);
-                    out.push(bir::Statement {
-                        kind: bir::StatementKind::Expr { value },
-                        span,
-                    });
-                }
-            }
-            ast::Statement::Assert(assert_stmt) => self.lower_assert(assert_stmt, scope, span, out),
-            ast::Statement::Pass => {}
-            ast::Statement::Break(value) => self.lower_break(value.as_ref(), scope, span, out),
-            ast::Statement::Continue => out.push(bir::Statement {
-                kind: bir::StatementKind::Continue,
-                span,
-            }),
-            other => self.push_unsupported_stmt(unsupported_stmt_label(other), span, out),
-        }
-    }
-
-    /// Lower an inferred/`let`/`mutable`/reassignment statement. A `Reassign` binding reuses the existing local for
-    /// `assignment.name` when one is already bound (falling back to declaring a new one if reassignment targets an
-    /// unbound name), while every other binding kind always declares a fresh local — matching source-level shadowing
-    /// semantics, where a repeated `let x = ...` introduces a new binding rather than mutating the old one.
-    fn lower_assignment(
-        &mut self,
-        assignment: &ast::AssignmentStmt,
-        remaining: &[ast::Spanned<ast::Statement>],
-        scope: bir::ScopeId,
-        span: HirSourceSpan,
-        out: &mut Vec<bir::Statement>,
-    ) {
-        // A closure value already carries the typechecker's callable shape. A partial retains that full callable
-        // type, with captured presets represented as named-overrideable defaults. Positional calls skip those preset
-        // slots, and `LocalCallableTarget::binding` records the resulting declaration mapping. Keeping this type on
-        // the binding makes the local call contract agree with the `Rvalue::Closure` that creates the value.
-        let ty = self
-            .callable_value_ty(&assignment.value)
-            .unwrap_or_else(|| self.resolve_ty(assignment.value.span));
-        let value = self.lower_expr_to_operand(&assignment.value, scope, out);
-        let local = match assignment.binding {
-            ast::BindingKind::Reassign => self
-                .bindings
-                .get(&assignment.name)
-                .copied()
-                .unwrap_or_else(|| self.declare_new_local(assignment.name.clone(), ty, scope, span, remaining)),
-            ast::BindingKind::Inferred | ast::BindingKind::Let | ast::BindingKind::Mutable => {
-                self.declare_new_local(assignment.name.clone(), ty, scope, span, remaining)
-            }
-        };
-        out.push(bir::Statement {
-            kind: bir::StatementKind::Assign {
-                place: bir::Place::from_local(local),
-                rvalue: bir::Rvalue::Use(value),
-            },
-            span,
-        });
-    }
-
-    /// Lower `obj.field = value` (including the compound `obj.field <op>= value` form). The parser already
-    /// desugars a compound `FieldAssignmentStmt` so `value` is the full `obj.field <op> rhs` expression
-    /// (`crates/incan_syntax/src/parser/stmts.rs`'s `assignment_or_expr_stmt`) -- `fa.compound_op` is purely a
-    /// formatter hint for round-tripping `+=` spelling and carries no separate lowering semantics here, so this
-    /// only needs to build the write-side place and lower `value` normally.
-    fn lower_field_assignment(
-        &mut self,
-        field_assignment: &ast::FieldAssignmentStmt,
-        scope: bir::ScopeId,
-        span: HirSourceSpan,
-        out: &mut Vec<bir::Statement>,
-    ) {
-        let mut place = self.lower_expr_to_place(&field_assignment.object, scope, out);
-        place
-            .projection
-            .push(bir::PlaceElem::Field(field_assignment.field.clone()));
-        let value = self.lower_expr_to_operand(&field_assignment.value, scope, out);
-        out.push(bir::Statement {
-            kind: bir::StatementKind::Assign {
-                place,
-                rvalue: bir::Rvalue::Use(value),
-            },
-            span,
-        });
-    }
-
-    /// Lower `obj[index] = value` (including the compound `obj[index] <op>= value` form, pre-desugared into
-    /// `value` by the parser -- see [`Self::lower_field_assignment`]'s docs for the same note on
-    /// `IndexAssignmentStmt::compound_op`). The object place is lowered before the index operand, preserving the
-    /// established assignment evaluation order in the Rust-emission backend: object, index, then assigned value.
-    fn lower_index_assignment(
-        &mut self,
-        index_assignment: &ast::IndexAssignmentStmt,
-        scope: bir::ScopeId,
-        span: HirSourceSpan,
-        out: &mut Vec<bir::Statement>,
-    ) {
-        let mut place = self.lower_expr_to_place(&index_assignment.object, scope, out);
-        let index_operand = self.lower_expr_to_operand(&index_assignment.index, scope, out);
-        place.projection.push(bir::PlaceElem::Index(Box::new(index_operand)));
-        let value = self.lower_expr_to_operand(&index_assignment.value, scope, out);
-        out.push(bir::Statement {
-            kind: bir::StatementKind::Assign {
-                place,
-                rvalue: bir::Rvalue::Use(value),
-            },
-            span,
-        });
-    }
-
-    /// Lower `name <op>= value` (`x += y`, `x &= y`, ...). Unlike field/index compound assignment, the parser
-    /// leaves `ca.value` as the plain right-hand operand rather than pre-desugaring it, so this explicitly reads
-    /// `name`'s current value, combines it with `value` via [`Self::lower_binary_from_operands`] (shared with
-    /// [`Self::lower_binary`], so string-concat compound assignment routes through the same helper-call machinery
-    /// as `+`), and writes the result back. An operator with no Body IR equivalent (see [`lower_binary_op`]) or a
-    /// name that is not currently bound (should not happen after a successful typecheck) falls back to an explicit
-    /// unsupported placeholder instead of panicking.
-    fn lower_compound_assignment(
-        &mut self,
-        compound_assignment: &ast::CompoundAssignmentStmt,
-        scope: bir::ScopeId,
-        span: HirSourceSpan,
-        out: &mut Vec<bir::Statement>,
-    ) {
-        let Some(&local) = self.bindings.get(&compound_assignment.name) else {
-            self.push_unsupported_stmt(
-                format!("compound assignment to unbound name `{}`", compound_assignment.name),
-                span,
-                out,
-            );
-            return;
-        };
-        let lhs_ty = self.locals[local.index()].ty.clone();
-        let op = compound_assignment.op.binary_op();
-        let rhs_ty = self.resolve_ty(compound_assignment.value.span);
-        if !Self::binary_op_is_supported(op, &lhs_ty, &rhs_ty) {
-            self.push_unsupported_stmt(
-                format!("compound assignment operator {:?}", compound_assignment.op),
-                span,
-                out,
-            );
-            return;
-        }
-        let lhs_place = bir::Place::from_local(local);
-        let (fact, last_use) = self.ownership_fact_for_place(&lhs_place, &lhs_ty);
-        let lhs_operand = bir::Operand::place(lhs_place, fact, last_use);
-        let rhs_operand = self.lower_expr_to_operand(&compound_assignment.value, scope, out);
-        let result = self.lower_binary_from_operands(
-            op,
-            &lhs_ty,
-            lhs_operand,
-            &rhs_ty,
-            rhs_operand,
-            lhs_ty.clone(),
-            scope,
-            span,
-            out,
-        );
-        out.push(bir::Statement {
-            kind: bir::StatementKind::Assign {
-                place: bir::Place::from_local(local),
-                rvalue: bir::Rvalue::Use(result),
-            },
-            span,
-        });
-    }
-
-    /// Resolve or declare the local for one name bound by a multi-target assignment (tuple unpack or chained
-    /// assignment). A `Reassign` binding reuses an existing local exactly like [`Self::lower_assignment`] does for
-    /// a plain single-target reassignment; every other binding kind always declares a fresh local, matching
-    /// source-level shadowing semantics.
-    fn bind_multi_target_name(
-        &mut self,
-        name: &str,
-        ty: IncanType,
-        binding: ast::BindingKind,
-        scope: bir::ScopeId,
-        span: HirSourceSpan,
-        remaining: &[ast::Spanned<ast::Statement>],
-    ) -> bir::LocalId {
-        match binding {
-            ast::BindingKind::Reassign => self
-                .bindings
-                .get(name)
-                .copied()
-                .unwrap_or_else(|| self.declare_new_local(name.to_string(), ty, scope, span, remaining)),
-            ast::BindingKind::Inferred | ast::BindingKind::Let | ast::BindingKind::Mutable => {
-                self.declare_new_local(name.to_string(), ty, scope, span, remaining)
-            }
-        }
-    }
-
-    /// Lower `a, b = value` / `let a, b = value` into a sequence of single-target `Assign` statements: materialize
-    /// `value` once, then bind each name to the corresponding `.{index}` tuple-field projection off it, in
-    /// left-to-right order. Element reads go through the same [`Self::ownership_fact_for_place`] a plain
-    /// `.field`/`[index]` read anywhere else in v0 uses, so a non-Copy element borrows rather than moves (v0 does
-    /// not track partial-move state out of a place, per [`Self::ownership_fact_for_place`]'s own docs) --
-    /// consistent with, not a special case of, that existing policy.
-    fn lower_tuple_unpack(
-        &mut self,
-        tuple_unpack: &ast::TupleUnpackStmt,
-        remaining: &[ast::Spanned<ast::Statement>],
-        scope: bir::ScopeId,
-        span: HirSourceSpan,
-        out: &mut Vec<bir::Statement>,
-    ) {
-        let value_ty = self.resolve_ty(tuple_unpack.value.span);
-        if let Some(reason) = unsupported_tuple_destructure(&value_ty, tuple_unpack.names.len()) {
-            self.push_unsupported_stmt(reason, span, out);
-            return;
-        }
-        let value_operand = self.lower_expr_to_operand(&tuple_unpack.value, scope, out);
-        let value_place = self.materialize_operand_to_place(value_operand, value_ty.clone(), scope, span, out);
-        let element_types = tuple_element_types(&value_ty, tuple_unpack.names.len());
-
-        for (index, (name, element_ty)) in tuple_unpack.names.iter().zip(&element_types).enumerate() {
-            let mut element_place = value_place.clone();
-            element_place.projection.push(bir::PlaceElem::Field(index.to_string()));
-            let (fact, last_use) = self.ownership_fact_for_place(&element_place, element_ty);
-            let element_operand = bir::Operand::place(element_place, fact, last_use);
-            let local =
-                self.bind_multi_target_name(name, element_ty.clone(), tuple_unpack.binding, scope, span, remaining);
-            out.push(bir::Statement {
-                kind: bir::StatementKind::Assign {
-                    place: bir::Place::from_local(local),
-                    rvalue: bir::Rvalue::Use(element_operand),
-                },
-                span,
-            });
-        }
-    }
-
-    /// Lower `t1, t2 = value` where the targets are lvalue expressions (`arr[i], arr[j] = ...`), not new bindings
-    /// -- used for swaps and other multi-target reassignments. Materializes `value` once, then reads and
-    /// materializes each element into its own fresh temporary *before* writing to any target, so aliased targets
-    /// and sources (for example `arr[i], arr[j] = arr[j], arr[i]`) read the pre-assignment values rather than one
-    /// another's already-written results. This is genuinely new coverage: the existing Rust-emission backend does
-    /// not implement `TupleAssign` at all (`src/backend/ir/lower/stmt.rs` returns a `LoweringError`), so there is
-    /// no existing behavior to mirror here -- the evaluation order above is v0's own design, chosen specifically
-    /// to make `a, b = b, a` swap correctly.
-    fn lower_tuple_assign(
-        &mut self,
-        tuple_assign: &ast::TupleAssignStmt,
-        scope: bir::ScopeId,
-        span: HirSourceSpan,
-        out: &mut Vec<bir::Statement>,
-    ) {
-        let value_ty = self.resolve_ty(tuple_assign.value.span);
-        if let Some(reason) = unsupported_tuple_destructure(&value_ty, tuple_assign.targets.len()) {
-            self.push_unsupported_stmt(reason, span, out);
-            return;
-        }
-        let value_operand = self.lower_expr_to_operand(&tuple_assign.value, scope, out);
-        let value_place = self.materialize_operand_to_place(value_operand, value_ty.clone(), scope, span, out);
-        let element_types = tuple_element_types(&value_ty, tuple_assign.targets.len());
-
-        let mut element_operands = Vec::with_capacity(tuple_assign.targets.len());
-        for (index, element_ty) in element_types.iter().enumerate() {
-            let mut element_place = value_place.clone();
-            element_place.projection.push(bir::PlaceElem::Field(index.to_string()));
-            let (fact, last_use) = self.ownership_fact_for_place(&element_place, element_ty);
-            let element_operand = bir::Operand::place(element_place, fact, last_use);
-            element_operands.push(self.push_assign_temp(
-                bir::Rvalue::Use(element_operand),
-                element_ty.clone(),
-                scope,
-                span,
-                out,
-            ));
-        }
-
-        for (target, value) in tuple_assign.targets.iter().zip(element_operands) {
-            let place = self.lower_expr_to_place(target, scope, out);
-            out.push(bir::Statement {
-                kind: bir::StatementKind::Assign {
-                    place,
-                    rvalue: bir::Rvalue::Use(value),
-                },
-                span,
-            });
-        }
-    }
-
-    /// Lower `x = y = z = value` into `z = value; y = <read z>; x = <read y>` (rightmost target first), matching
-    /// the direction the existing Rust-emission backend already chose for this same desugar
-    /// (`src/backend/ir/lower/stmt.rs`'s `ChainedAssignment` arm).
-    fn lower_chained_assignment(
-        &mut self,
-        chained_assignment: &ast::ChainedAssignmentStmt,
-        remaining: &[ast::Spanned<ast::Statement>],
-        scope: bir::ScopeId,
-        span: HirSourceSpan,
-        out: &mut Vec<bir::Statement>,
-    ) {
-        let Some(last_name) = chained_assignment.targets.last() else {
-            self.push_unsupported_stmt("empty chained assignment".to_string(), span, out);
-            return;
-        };
-        let value_ty = self.resolve_ty(chained_assignment.value.span);
-        let value_operand = self.lower_expr_to_operand(&chained_assignment.value, scope, out);
-        let mut prev_local = self.bind_multi_target_name(
-            last_name,
-            value_ty.clone(),
-            chained_assignment.binding,
-            scope,
-            span,
-            remaining,
-        );
-        out.push(bir::Statement {
-            kind: bir::StatementKind::Assign {
-                place: bir::Place::from_local(prev_local),
-                rvalue: bir::Rvalue::Use(value_operand),
-            },
-            span,
-        });
-
-        // Walk the remaining targets right-to-left, each one reading the local immediately to its right.
-        for name in chained_assignment.targets[..chained_assignment.targets.len() - 1]
-            .iter()
-            .rev()
-        {
-            // `remaining_reads[prev_local]` was seeded only from statements *after* this whole chained-assignment
-            // statement (see `Self::declare_new_local`'s `remaining` parameter) -- it does not know about the
-            // synthetic read performed right here, within the very statement that (re)bound `prev_local`. Bump it
-            // by one first so the shared `Self::ownership_fact_for_place` decrement below still lands on the
-            // correct move/clone decision instead of under-counting by one.
-            if let Some(remaining_count) = self.remaining_reads.get_mut(&prev_local) {
-                *remaining_count += 1;
-            }
-            let place = bir::Place::from_local(prev_local);
-            let (fact, last_use) = self.ownership_fact_for_place(&place, &value_ty);
-            let operand = bir::Operand::place(place, fact, last_use);
-            let local = self.bind_multi_target_name(
-                name,
-                value_ty.clone(),
-                chained_assignment.binding,
-                scope,
-                span,
-                remaining,
-            );
-            out.push(bir::Statement {
-                kind: bir::StatementKind::Assign {
-                    place: bir::Place::from_local(local),
-                    rvalue: bir::Rvalue::Use(operand),
-                },
-                span,
-            });
-            prev_local = local;
-        }
-    }
-
-    /// Lower a `break` / `break value` statement. A value routes into the innermost enclosing loop's result place
-    /// when that loop is a value-producing `loop:` expression (see [`Self::lower_loop_expr`]) -- otherwise it stays
-    /// on the `Break` statement itself, matching [`bir::StatementKind::Break`]'s documented default. The innermost
-    /// context comes from [`Self::loop_break_targets`], which every loop-lowering path pushes/pops around its own
-    /// body so a `break` always targets the loop it is lexically inside, never an outer one.
-    fn lower_break(
-        &mut self,
-        value: Option<&ast::Spanned<ast::Expr>>,
-        scope: bir::ScopeId,
-        span: HirSourceSpan,
-        out: &mut Vec<bir::Statement>,
-    ) {
-        let target = self.loop_break_targets.last().copied().flatten();
-        match (value, target) {
-            (Some(expr), Some(result_local)) => {
-                let operand = self.lower_expr_to_operand(expr, scope, out);
-                out.push(bir::Statement {
-                    kind: bir::StatementKind::Assign {
-                        place: bir::Place::from_local(result_local),
-                        rvalue: bir::Rvalue::Use(operand),
-                    },
-                    span,
-                });
-                out.push(bir::Statement {
-                    kind: bir::StatementKind::Break { value: None },
-                    span,
-                });
-            }
-            _ => {
-                let operand = value.map(|v| self.lower_expr_to_operand(v, scope, out));
-                out.push(bir::Statement {
-                    kind: bir::StatementKind::Break { value: operand },
-                    span,
-                });
-            }
-        }
-    }
-
-    /// Lower a statement-position `yield value` (`ast::Expr::Yield(Some(value))` reached through
-    /// [`Self::lower_stmt_into`]'s `ast::Statement::Expr` arm) into a [`bir::StatementKind::Yield`].
-    ///
-    /// `value` is lowered through the same [`Self::lower_expr_to_operand`] path every other statement's operand
-    /// goes through, so ownership facts/last-use tracking apply to a yielded value exactly like any other read.
-    /// Records the runtime dependencies the existing Rust-emission backend's own `yield` lowering actually needs
-    /// (`__incan_yield.yield_value(..)` on a `GeneratorYield` handle backed by `std::thread::spawn` and
-    /// `std::sync::mpsc::sync_channel` -- see `crates/incan_stdlib/src/iter.rs`'s `Generator`/`SpawnedGenerator`):
-    /// a named runtime helper (mirroring how [`Self::lower_fstring`] records `"fstring"` without a new
-    /// [`bir::HelperOp`] variant, since `Yield` is its own statement kind, not a [`bir::Callee::Helper`] call),
-    /// [`AbiV0RuntimeRequirement::HostedStd`] (the spawned-thread/channel machinery is not freestanding-compatible),
-    /// and [`AbiV0RuntimeRequirement::Allocator`] (the channel and boxed iterator both allocate).
-    fn lower_yield(
-        &mut self,
-        value: &ast::Spanned<ast::Expr>,
-        scope: bir::ScopeId,
-        span: HirSourceSpan,
-        out: &mut Vec<bir::Statement>,
-    ) {
-        let operand = self.lower_expr_to_operand(value, scope, out);
-        out.push(bir::Statement {
-            kind: bir::StatementKind::Yield { value: operand },
-            span,
-        });
-        self.record_runtime_requirement(AbiV0RuntimeRequirement::RuntimeHelper("generator".to_string()));
-        self.record_runtime_requirement(AbiV0RuntimeRequirement::HostedStd);
-        self.record_runtime_requirement(AbiV0RuntimeRequirement::Allocator);
-    }
-
-    /// Lower `if`/`elif`/`else` into a [`bir::StatementKind::If`] chain. `elif` branches are folded into nested
-    /// `else { if ... }` wrappers from the last branch inward (see the inline comment above the fold loop), and an
-    /// `if let` pattern condition — not yet modeled by v0 — lowers to an explicit unsupported placeholder instead of
-    /// the real branch.
-    fn lower_if(
-        &mut self,
-        if_stmt: &ast::IfStmt,
-        scope: bir::ScopeId,
-        span: HirSourceSpan,
-        out: &mut Vec<bir::Statement>,
-    ) {
-        let ast::Condition::Expr(cond_expr) = &if_stmt.condition else {
-            self.push_unsupported_stmt("if-let pattern condition".to_string(), span, out);
-            return;
-        };
-        let cond = self.lower_expr_to_operand(cond_expr, scope, out);
-
-        let then_block = self.lower_branch_block(&if_stmt.then_body, scope, span);
-        let mut else_block = if_stmt
-            .else_body
-            .as_ref()
-            .map(|else_body| self.lower_branch_block(else_body, scope, span));
-
-        // Fold `elif` branches into nested `else { if ... }` wrappers, innermost (last elif) first, so the earlier
-        // conditions end up evaluated first at the top of the chain once wrapped by the outer `if` pushed below.
-        for (elif_cond, elif_body) in if_stmt.elif_branches.iter().rev() {
-            let mut wrapper = Vec::new();
-            let cond_operand = self.lower_expr_to_operand(elif_cond, scope, &mut wrapper);
-            let then_block = self.lower_branch_block(elif_body, scope, span);
-            wrapper.push(bir::Statement {
-                kind: bir::StatementKind::If {
-                    cond: cond_operand,
-                    then_block,
-                    else_block,
-                },
-                span,
-            });
-            else_block = Some(bir::Block { scope, stmts: wrapper });
-        }
-
-        out.push(bir::Statement {
-            kind: bir::StatementKind::If {
-                cond,
-                then_block,
-                else_block,
-            },
-            span,
-        });
-    }
-
-    /// Lower one `if`/`elif`/`else` branch body into its own scoped [`bir::Block`]: allocate a child scope, lower
-    /// the statements into it, then insert scope-exit drops. Shared by [`Self::lower_if`]'s then/else/elif bodies
-    /// and [`Self::lower_if_expr`]'s then/else bodies, since both need exactly this shape.
-    fn lower_branch_block(
-        &mut self,
-        body: &[ast::Spanned<ast::Statement>],
-        parent_scope: bir::ScopeId,
-        span: HirSourceSpan,
-    ) -> bir::Block {
-        let branch_scope = self.new_scope(Some(parent_scope), span);
-        let mut stmts = Vec::new();
-        self.lower_block_into(body, branch_scope, &mut stmts);
-        self.insert_scope_drops(&mut stmts, branch_scope);
-        bir::Block {
-            scope: branch_scope,
-            stmts,
-        }
-    }
-
-    /// Lower an expression-position `if` (`ast::Expr::If`) into the same [`bir::StatementKind::If`] shape
-    /// statement-position `if` uses (see [`Self::lower_if`]), reusing [`Self::lower_branch_block`] for both
-    /// branches. The typechecker gives an expression-position `if` type `Unit` unconditionally (`check_if_expr` in
-    /// `src/frontend/typechecker/check_expr/control_flow.rs` discards any branch value and always returns
-    /// `ResolvedType::Unit`) -- unlike a `loop` expression, an `if` expression cannot yet produce a value from its
-    /// branches, so its Body IR operand is always the `Unit` constant rather than a place read.
-    fn lower_if_expr(
-        &mut self,
-        if_expr: &ast::IfExpr,
-        scope: bir::ScopeId,
-        span: ast::Span,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let hir_span_value = hir_span(span);
-        let cond = self.lower_expr_to_operand(&if_expr.condition, scope, out);
-        let then_block = self.lower_branch_block(&if_expr.then_body, scope, hir_span_value);
-        let else_block = if_expr
-            .else_body
-            .as_ref()
-            .map(|body| self.lower_branch_block(body, scope, hir_span_value));
-        out.push(bir::Statement {
-            kind: bir::StatementKind::If {
-                cond,
-                then_block,
-                else_block,
-            },
-            span: hir_span_value,
-        });
-        bir::Operand::Constant(bir::Constant::Unit)
-    }
-
-    /// Lower a value-producing `loop:` expression (`ast::Expr::Loop`) into a [`bir::StatementKind::Loop`] plus a
-    /// dedicated result local that every `break value` inside the loop's *own* body (not a nested loop's --
-    /// enforced by [`Self::loop_break_targets`]) assigns into before exiting. The typechecker resolves this
-    /// expression's type from the union of its `break value` operand types (`check_loop_expr` in
-    /// `src/frontend/typechecker/check_expr/control_flow.rs`), so -- unlike an `if` expression, which is always
-    /// `Unit` -- a `loop` expression's produced value genuinely comes from its branches and needs this
-    /// merge-into-one-place treatment; see [`Self::lower_break`] for the other half of the mechanism.
-    fn lower_loop_expr(
-        &mut self,
-        loop_expr: &ast::LoopExpr,
-        scope: bir::ScopeId,
-        span: ast::Span,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let hir_span_value = hir_span(span);
-        let ty = self.resolve_ty(span);
-        let loop_scope = self.new_scope(Some(scope), hir_span_value);
-        let result_local = self.new_temp(ty.clone(), loop_scope, hir_span_value);
-
-        self.loop_break_targets.push(Some(result_local));
-        let mut body_stmts = Vec::new();
-        self.lower_block_into(&loop_expr.body, loop_scope, &mut body_stmts);
-        self.insert_scope_drops(&mut body_stmts, loop_scope);
-        self.loop_break_targets.pop();
-
-        out.push(bir::Statement {
-            kind: bir::StatementKind::Loop {
-                body: bir::Block {
-                    scope: loop_scope,
-                    stmts: body_stmts,
-                },
-            },
-            span: hir_span_value,
-        });
-        self.temp_operand(result_local, &ty)
-    }
-
-    /// Lower `while cond: body` into Body IR's single normalized loop shape: a [`bir::StatementKind::Loop`] whose
-    /// body opens with `if not cond: break`, followed by the lowered loop body. A `while let` pattern condition —
-    /// not yet modeled by v0 — lowers to an explicit unsupported placeholder instead of the real loop.
-    fn lower_while(
-        &mut self,
-        while_stmt: &ast::WhileStmt,
-        scope: bir::ScopeId,
-        span: HirSourceSpan,
-        out: &mut Vec<bir::Statement>,
-    ) {
-        let ast::Condition::Expr(cond_expr) = &while_stmt.condition else {
-            self.push_unsupported_stmt("while-let pattern condition".to_string(), span, out);
-            return;
-        };
-
-        let loop_scope = self.new_scope(Some(scope), span);
-        // `while` never produces a value from `break`, so push `None`: a `break` inside this loop's body must
-        // resolve to a plain valueless exit even if this `while` is lexically nested inside a value-producing
-        // `loop:` expression (see `Self::loop_break_targets`'s docs for why the stack exists).
-        self.loop_break_targets.push(None);
-        let mut body_stmts = Vec::new();
-        let cond_operand = self.lower_expr_to_operand(cond_expr, loop_scope, &mut body_stmts);
-        let negated = self.negate_operand(cond_operand, loop_scope, span, &mut body_stmts);
-        let break_scope = self.new_scope(Some(loop_scope), span);
-        let break_block = bir::Block {
-            scope: break_scope,
-            stmts: vec![bir::Statement {
-                kind: bir::StatementKind::Break { value: None },
-                span,
-            }],
-        };
-        body_stmts.push(bir::Statement {
-            kind: bir::StatementKind::If {
-                cond: negated,
-                then_block: break_block,
-                else_block: None,
-            },
-            span,
-        });
-
-        self.lower_block_into(&while_stmt.body, loop_scope, &mut body_stmts);
-        self.insert_scope_drops(&mut body_stmts, loop_scope);
-        self.loop_break_targets.pop();
-
-        out.push(bir::Statement {
-            kind: bir::StatementKind::Loop {
-                body: bir::Block {
-                    scope: loop_scope,
-                    stmts: body_stmts,
-                },
-            },
-            span,
-        });
-    }
-
-    /// Lower a `for` statement. `for x in start..end: body` (range-shaped iterables) lowers into a normalized
-    /// counting `Loop`, preserving #1103's original range-loop shape unchanged. Every other iterable -- builtin
-    /// collections (`List`/`Dict`/`String`) and user-defined iterables implementing the RFC 068 `__iter__`/
-    /// `__next__` protocol, including the fallible `for item in iterable?:` form (RFC 115) -- lowers through
-    /// [`Self::lower_general_iteration`], sharing its per-clause iteration primitive with comprehensions and
-    /// generator expressions (see [`Self::lower_comprehension_clauses`]).
-    ///
-    /// Both paths accept the same loop-pattern subset the typechecker accepts -- a plain binding, `_`, and
-    /// (recursively) a tuple of those, per `TypeChecker::define_for_pattern_bindings` in
-    /// `src/frontend/typechecker/check_stmt.rs` (#1125). A plain `for x in ...` binds the produced item directly;
-    /// every other shape writes it into a per-iteration temporary that [`Self::bind_for_pattern`] then projects one
-    /// real named binding out of per bound name. Any shape outside that subset -- which the typechecker already
-    /// rejects with its own diagnostic before lowering ever runs -- lowers to `Unsupported` naming the offending
-    /// shape, checked up front so a refusal never leaves half-emitted bindings behind (the same
-    /// "check before partially lowering" precedent as [`Self::lower_binary`] and [`Self::lower_match`]). The same
-    /// up-front check also refuses a tuple pattern whose produced item is not a tuple of matching arity, so
-    /// lowering can never invent `.0`/`.1` projections into a value that has no such fields -- see
-    /// [`unsupported_for_pattern`].
-    fn lower_for(
-        &mut self,
-        for_stmt: &ast::ForStmt,
-        scope: bir::ScopeId,
-        span: HirSourceSpan,
-        out: &mut Vec<bir::Statement>,
-    ) {
-        let item_ty = self.resolve_ty(for_stmt.pattern.span);
-        if let Some(reason) = unsupported_for_pattern(&for_stmt.pattern.node, &item_ty) {
-            self.push_unsupported_stmt(reason, span, out);
-            return;
-        }
-        // The typechecker enters a lexical block scope for the loop header/body, so every binding introduced by the
-        // pattern must disappear after the statement. Keep the active lookup map for restoration while leaving the
-        // loop locals themselves in Body IR for the loop's statements to reference.
-        let enclosing_bindings = self.bindings.clone();
-        let ast::Expr::Range { start, end, inclusive } = &for_stmt.iter.node else {
-            let loop_scope = self.new_scope(Some(scope), span);
-            let item_local = self.declare_for_item_local(&for_stmt.pattern, &item_ty, loop_scope, span, &for_stmt.body);
-            self.lower_general_iteration(
-                &for_stmt.iter,
-                item_local,
-                scope,
-                loop_scope,
-                span,
-                out,
-                |builder, loop_scope, body_stmts| {
-                    builder.bind_for_pattern(
-                        &for_stmt.pattern,
-                        &item_ty,
-                        item_local,
-                        loop_scope,
-                        &for_stmt.body,
-                        body_stmts,
-                    );
-                    builder.lower_block_into(&for_stmt.body, loop_scope, body_stmts);
-                    builder.insert_scope_drops(body_stmts, loop_scope);
-                },
-            );
-            self.bindings = enclosing_bindings;
-            return;
-        };
-
-        let int_ty = IncanType::Primitive(IncanPrimitiveType::Int);
-        let start_operand = self.lower_expr_to_operand(start, scope, out);
-        let idx_local = self.new_temp(int_ty.clone(), scope, span);
-        out.push(bir::Statement {
-            kind: bir::StatementKind::Assign {
-                place: bir::Place::from_local(idx_local),
-                rvalue: bir::Rvalue::Use(start_operand),
-            },
-            span,
-        });
-
-        let loop_scope = self.new_scope(Some(scope), span);
-        // `for` never produces a value from `break` (same reasoning as `while` -- see `Self::lower_while`).
-        self.loop_break_targets.push(None);
-        let mut body_stmts = Vec::new();
-
-        let end_operand = self.lower_expr_to_operand(end, loop_scope, &mut body_stmts);
-        let idx_read = bir::Operand::place(bir::Place::from_local(idx_local), bir::OwnershipFact::Copy, false);
-        let cmp_op = if *inclusive { bir::BinOp::Gt } else { bir::BinOp::Ge };
-        let cond = self.push_assign_temp(
-            bir::Rvalue::BinaryOp(cmp_op, idx_read, end_operand),
-            IncanType::Primitive(IncanPrimitiveType::Bool),
-            loop_scope,
-            span,
-            &mut body_stmts,
-        );
-        let break_scope = self.new_scope(Some(loop_scope), span);
-        body_stmts.push(bir::Statement {
-            kind: bir::StatementKind::If {
-                cond,
-                then_block: bir::Block {
-                    scope: break_scope,
-                    stmts: vec![bir::Statement {
-                        kind: bir::StatementKind::Break { value: None },
-                        span,
-                    }],
-                },
-                else_block: None,
-            },
-            span,
-        });
-
-        // `for _ in start..end` binds nothing and the range's own index already drives the loop, so it needs no
-        // per-iteration item local at all -- unlike the general path, where `IterNext` must still write the polled
-        // item somewhere for the poll itself to happen.
-        if !matches!(for_stmt.pattern.node, ast::Pattern::Wildcard) {
-            let item_local = self.declare_for_item_local(&for_stmt.pattern, &item_ty, loop_scope, span, &for_stmt.body);
-            body_stmts.push(bir::Statement {
-                kind: bir::StatementKind::Assign {
-                    place: bir::Place::from_local(item_local),
-                    rvalue: bir::Rvalue::Use(bir::Operand::place(
-                        bir::Place::from_local(idx_local),
-                        bir::OwnershipFact::Copy,
-                        false,
-                    )),
-                },
-                span,
-            });
-            self.bind_for_pattern(
-                &for_stmt.pattern,
-                &item_ty,
-                item_local,
-                loop_scope,
-                &for_stmt.body,
-                &mut body_stmts,
-            );
-        }
-
-        self.lower_block_into(&for_stmt.body, loop_scope, &mut body_stmts);
-        self.insert_scope_drops(&mut body_stmts, loop_scope);
-
-        let one = bir::Operand::Constant(bir::Constant::Int(1));
-        let idx_read_for_incr = bir::Operand::place(bir::Place::from_local(idx_local), bir::OwnershipFact::Copy, false);
-        let incremented = self.push_assign_temp(
-            bir::Rvalue::BinaryOp(bir::BinOp::Add, idx_read_for_incr, one),
-            int_ty,
-            loop_scope,
-            span,
-            &mut body_stmts,
-        );
-        body_stmts.push(bir::Statement {
-            kind: bir::StatementKind::Assign {
-                place: bir::Place::from_local(idx_local),
-                rvalue: bir::Rvalue::Use(incremented),
-            },
-            span,
-        });
-        self.loop_break_targets.pop();
-
-        out.push(bir::Statement {
-            kind: bir::StatementKind::Loop {
-                body: bir::Block {
-                    scope: loop_scope,
-                    stmts: body_stmts,
-                },
-            },
-            span,
-        });
-        self.bindings = enclosing_bindings;
-    }
-
-    /// Declare the local each produced item of a `for` loop is written into.
-    ///
-    /// A plain `for x in ...` binds the item directly: the item local *is* `x`'s local, so the produced value is
-    /// never copied and the loop shape #1103/#1101 established is preserved byte-for-byte. Every other supported
-    /// pattern shape has no single name to write into, so the item goes into a temporary that
-    /// [`Self::bind_for_pattern`] projects the real bindings out of -- the same "materialize once, then bind each
-    /// element off a projection" shape [`Self::lower_tuple_unpack`] already uses for `a, b = value`.
-    fn declare_for_item_local(
-        &mut self,
-        pattern: &ast::Spanned<ast::Pattern>,
-        item_ty: &IncanType,
-        loop_scope: bir::ScopeId,
-        span: HirSourceSpan,
-        body: &[ast::Spanned<ast::Statement>],
-    ) -> bir::LocalId {
-        match &pattern.node {
-            ast::Pattern::Binding(name) => {
-                let total_reads = count_reads_in_stmts(name, body);
-                self.declare_new_local_with_reads(name.clone(), item_ty.clone(), loop_scope, span, total_reads)
-            }
-            _ => self.new_temp(item_ty.clone(), loop_scope, span),
-        }
-    }
-
-    /// Emit the binding statements a `for` loop's pattern needs against the item local, immediately after the
-    /// per-iteration `IterNext` (or, on the range path, after the index copy) has written it.
-    ///
-    /// A bare [`ast::Pattern::Binding`] emits nothing: [`Self::declare_for_item_local`] already declared the item
-    /// local *as* that binding, so there is nothing left to project. Every other shape delegates to
-    /// [`Self::bind_for_pattern_fields`], which means every binding that walk reaches is nested under at least one
-    /// tuple field and therefore always reads through a projection.
-    fn bind_for_pattern(
-        &mut self,
-        pattern: &ast::Spanned<ast::Pattern>,
-        item_ty: &IncanType,
-        item_local: bir::LocalId,
-        loop_scope: bir::ScopeId,
-        body: &[ast::Spanned<ast::Statement>],
-        out: &mut Vec<bir::Statement>,
-    ) {
-        if matches!(pattern.node, ast::Pattern::Binding(_)) {
-            return;
-        }
-        let item_place = bir::Place::from_local(item_local);
-        self.bind_for_pattern_fields(pattern, item_ty, &item_place, loop_scope, body, out);
-    }
-
-    /// Recursively bind one `for`-pattern node against `place`, the (already projected) part of the produced item
-    /// it corresponds to, emitting one `Assign` per bound name in source order.
-    ///
-    /// Iteration binding is *irrefutable*: unlike [`Self::lower_match_pattern`], which builds a [`bir::Pattern`]
-    /// for match-arm dispatch, there is nothing here to test or branch on, so this walk emits plain assignments and
-    /// deliberately does not reuse that machinery (#1125 names conflating the two as a non-goal). What it does
-    /// share is that walk's projection convention -- the zero-based tuple-element index spelled as a
-    /// [`bir::PlaceElem::Field`], matching [`Self::lower_tuple_unpack`]'s `.0`/`.1` spelling -- and its
-    /// [`tuple_element_types`] source for per-element types, so a nested tuple keeps resolved element types all the
-    /// way down and falls back to [`IncanType::Unknown`] per slot only where the resolved type is not a tuple of
-    /// the right arity.
-    ///
-    /// Each element is read through [`Self::ownership_fact_for_place`], exactly as
-    /// [`Self::lower_tuple_unpack`] reads its own elements, so a non-Copy element borrows rather than moving out of
-    /// a place v0 does not track partial-move state for. Each bound name becomes a real
-    /// [`bir::LocalOrigin::UserBinding`] local in `loop_scope`, seeded with its own last-use countdown over the
-    /// loop body, so [`Self::insert_scope_drops`] gives every non-Copy binding an explicit per-iteration drop.
-    ///
-    /// [`unsupported_for_pattern`] has already rejected every shape outside the accepted subset -- and every item
-    /// type that is not a tuple of matching arity -- before [`Self::lower_for`] reaches this walk, so the remaining
-    /// arms are unreachable in practice; they emit nothing rather than panicking if that invariant is ever violated
-    /// by a hand-built AST.
-    fn bind_for_pattern_fields(
-        &mut self,
-        pattern: &ast::Spanned<ast::Pattern>,
-        expected_ty: &IncanType,
-        place: &bir::Place,
-        loop_scope: bir::ScopeId,
-        body: &[ast::Spanned<ast::Statement>],
-        out: &mut Vec<bir::Statement>,
-    ) {
-        let span = hir_span(pattern.span);
-        match &pattern.node {
-            ast::Pattern::Wildcard => {}
-            ast::Pattern::Binding(name) => {
-                let (fact, last_use) = self.ownership_fact_for_place(place, expected_ty);
-                let element = bir::Operand::place(place.clone(), fact, last_use);
-                let total_reads = count_reads_in_stmts(name, body);
-                let local =
-                    self.declare_new_local_with_reads(name.clone(), expected_ty.clone(), loop_scope, span, total_reads);
-                out.push(bir::Statement {
-                    kind: bir::StatementKind::Assign {
-                        place: bir::Place::from_local(local),
-                        rvalue: bir::Rvalue::Use(element),
-                    },
-                    span,
-                });
-            }
-            ast::Pattern::Tuple(items) => {
-                let element_types = tuple_element_types(expected_ty, items.len());
-                for (index, (item, element_ty)) in items.iter().zip(&element_types).enumerate() {
-                    let mut field_place = place.clone();
-                    field_place.projection.push(bir::PlaceElem::Field(index.to_string()));
-                    self.bind_for_pattern_fields(item, element_ty, &field_place, loop_scope, body, out);
-                }
-            }
-            ast::Pattern::Literal(_) | ast::Pattern::Constructor(..) | ast::Pattern::Group(_) | ast::Pattern::Or(_) => {
-            }
-        }
-    }
-
-    /// Lower one general (non-range) iteration: materialize an iterator from `iter_expr` before the loop, then push
-    /// a single [`bir::StatementKind::Loop`] whose body opens with a [`bir::StatementKind::IterNext`] writing each
-    /// produced item into `pattern_local`, followed by `body_fn`. Shared by [`Self::lower_for`]'s general-iterable
-    /// path and [`Self::lower_comprehension_clauses`]'s `for`-clause handling, so builtin-vs-protocol iteration is
-    /// resolved in exactly one place rather than twice.
-    ///
-    /// Looks up [`TypeCheckInfo::protocol_iteration`] at `iter_expr`'s span to decide the [`bir::IterProtocol`]:
-    /// `None` means a builtin collection or range, where "the iterator" is modeled as the iterable's own value (no
-    /// method dispatch) -- a plain `Assign`; `Some` means a resolved `__iter__`/`__next__` protocol, where the
-    /// iterator is obtained via an explicit `iter_method` [`bir::Callee::Method`] call. When the resolved protocol
-    /// is fallible (`for item in iterable?:`, RFC 115), `iter_expr` is itself `ast::Expr::Try(inner)` with the `?`
-    /// acting as the fallible-poll marker rather than an ordinary `Result` unwrap -- `inner` is lowered directly as
-    /// the iterable in that case (matching the existing Rust-emission backend's own `(Expr::Try(inner), Some(_)) =>
-    /// lower inner` special case in `src/backend/ir/lower/stmt.rs`), so the marker `?` is not double-lowered through
-    /// [`Self::lower_try`]. Any other `Expr::Try` (an ordinary `for item in result_of_iterable?:` unwrap) falls
-    /// through to the normal expression-lowering path, which already turns it into a
-    /// [`bir::StatementKind::TryPropagate`] ahead of the loop via [`Self::lower_expr_to_place`]'s existing
-    /// `Expr::Try` handling -- no special-casing needed for that form.
-    ///
-    /// The iterable is always read as a [`bir::OwnershipFact::Borrow`], matching
-    /// [`Self::lower_method_call`]'s established receiver-borrow precedent (never an unsound move, and consistent
-    /// with obtaining an iterator conceptually borrowing its source rather than consuming it at this normalized
-    /// level); the materialized iterator local is polled with [`bir::OwnershipFact::MutBorrow`] each iteration,
-    /// since polling advances its internal state.
-    #[allow(clippy::too_many_arguments)]
-    fn lower_general_iteration(
-        &mut self,
-        iter_expr: &ast::Spanned<ast::Expr>,
-        pattern_local: bir::LocalId,
-        outer_scope: bir::ScopeId,
-        loop_scope: bir::ScopeId,
-        span: HirSourceSpan,
-        out: &mut Vec<bir::Statement>,
-        body_fn: impl FnOnce(&mut Self, bir::ScopeId, &mut Vec<bir::Statement>),
-    ) {
-        let protocol = self.type_info.protocol_iteration(iter_expr.span).cloned();
-        let fallible = protocol.as_ref().is_some_and(|p| p.fallible_error_type.is_some());
-        let effective_iter_expr: &ast::Spanned<ast::Expr> = match (&iter_expr.node, fallible) {
-            (ast::Expr::Try(inner), true) => inner,
-            _ => iter_expr,
-        };
-
-        let iterable_place = self.lower_expr_to_place(effective_iter_expr, outer_scope, out);
-        let iterator_ty = match &protocol {
-            Some(p) => semantic_type_from_resolved(&p.iterator_type),
-            None => self.resolve_ty(effective_iter_expr.span),
-        };
-        let iterator_local = self.new_temp(iterator_ty, outer_scope, span);
-        match &protocol {
-            Some(p) => out.push(bir::Statement {
-                kind: bir::StatementKind::Call {
-                    destination: Some(bir::Place::from_local(iterator_local)),
-                    callee: bir::Callee::Method(bir::MethodTarget::synthesized(p.iter_method.clone())),
-                    args: fixed_elements(vec![bir::Operand::place(
-                        iterable_place,
-                        bir::OwnershipFact::Borrow,
-                        false,
-                    )]),
-                    may_panic: false,
-                },
-                span,
-            }),
-            None => out.push(bir::Statement {
-                kind: bir::StatementKind::Assign {
-                    place: bir::Place::from_local(iterator_local),
-                    rvalue: bir::Rvalue::Use(bir::Operand::place(iterable_place, bir::OwnershipFact::Borrow, false)),
-                },
-                span,
-            }),
-        }
-
-        self.loop_break_targets.push(None);
-        let mut body_stmts = Vec::new();
-
-        let iter_protocol = match &protocol {
-            Some(p) => bir::IterProtocol::UserDefined {
-                next_method: p.next_method.clone(),
-                fallible,
-            },
-            None => bir::IterProtocol::Builtin,
-        };
-        body_stmts.push(bir::Statement {
-            kind: bir::StatementKind::IterNext {
-                destination: bir::Place::from_local(pattern_local),
-                iterator: bir::Operand::place(
-                    bir::Place::from_local(iterator_local),
-                    bir::OwnershipFact::MutBorrow,
-                    false,
-                ),
-                protocol: iter_protocol,
-            },
-            span,
-        });
-
-        body_fn(self, loop_scope, &mut body_stmts);
-        self.loop_break_targets.pop();
-
-        out.push(bir::Statement {
-            kind: bir::StatementKind::Loop {
-                body: bir::Block {
-                    scope: loop_scope,
-                    stmts: body_stmts,
-                },
-            },
-            span,
-        });
-    }
-
-    /// Lower a list comprehension `[expr for pattern in iter if filter]` into: an empty
-    /// `AggregateKind::List` temporary, the desugared clause-chain loop (see
-    /// [`Self::lower_comprehension_clauses`]), pushing each accepted element into it via a compiler-synthesized
-    /// `push` [`bir::Callee::Method`] call, then a read of the completed list. Only v0's single mirrored
-    /// `(pattern, iter, filter)` clause is lowered -- `comp.clauses` is intentionally not consulted, since neither
-    /// the typechecker (`check_list_comp` in `src/frontend/typechecker/check_expr/comps.rs`) nor the existing
-    /// Rust-emission backend (`src/backend/ir/lower/expr/comprehensions.rs`) reads it either; a list comprehension
-    /// with more than one `for` clause is not actually type-checked or emitted as multi-clause today; treating
-    /// `comp.clauses` as authoritative here would silently lower a shape nothing else in the pipeline validates.
-    fn lower_list_comp(
-        &mut self,
-        comp: &ast::ListComp,
-        span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let hir_span_value = hir_span(span);
-        let ty = self.resolve_ty(span);
-        let list_local = self.new_temp(ty.clone(), scope, hir_span_value);
-        out.push(bir::Statement {
-            kind: bir::StatementKind::Assign {
-                place: bir::Place::from_local(list_local),
-                rvalue: bir::Rvalue::Aggregate(bir::AggregateKind::List, Vec::new()),
-            },
-            span: hir_span_value,
-        });
-        self.record_runtime_requirement(AbiV0RuntimeRequirement::Allocator);
-
-        let clauses = single_comprehension_clauses(&comp.pattern, &comp.iter, comp.filter.as_ref());
-        let terminal = ComprehensionTerminal::ListPush {
-            list_local,
-            element: &comp.expr,
-        };
-        self.lower_scoped_comprehension_clauses(&clauses, &terminal, scope, hir_span_value, out);
-        self.temp_operand(list_local, &ty)
-    }
-
-    /// Lower a dict comprehension `{key: value for pattern in iter if filter}` the same way
-    /// [`Self::lower_list_comp`] lowers a list comprehension, but growing an `AggregateKind::Dict` temporary via a
-    /// compiler-synthesized `insert` call. See [`Self::lower_list_comp`]'s docs for why only the single mirrored
-    /// clause is lowered, not `comp.clauses`.
-    fn lower_dict_comp(
-        &mut self,
-        comp: &ast::DictComp,
-        span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let hir_span_value = hir_span(span);
-        let ty = self.resolve_ty(span);
-        let dict_local = self.new_temp(ty.clone(), scope, hir_span_value);
-        out.push(bir::Statement {
-            kind: bir::StatementKind::Assign {
-                place: bir::Place::from_local(dict_local),
-                rvalue: bir::Rvalue::Dict(Vec::new()),
-            },
-            span: hir_span_value,
-        });
-        self.record_runtime_requirement(AbiV0RuntimeRequirement::Allocator);
-
-        let clauses = single_comprehension_clauses(&comp.pattern, &comp.iter, comp.filter.as_ref());
-        let terminal = ComprehensionTerminal::DictInsert {
-            dict_local,
-            key: &comp.key,
-            value: &comp.value,
-        };
-        self.lower_scoped_comprehension_clauses(&clauses, &terminal, scope, hir_span_value, out);
-        self.temp_operand(dict_local, &ty)
-    }
-
-    /// Lower a generator expression into a distinct, deferred [`bir::Rvalue::Generator`].
-    ///
-    /// The first `for` source is evaluated exactly once at construction, matching the established legacy
-    /// iterator-adapter emitter. Its value and every other needed outer lexical value are then captured into fresh
-    /// generator-local bindings. Clause polling, later `for` sources, filters, and element evaluation lower only
-    /// into the generator body, so the enclosing body neither materializes the sequence nor runs a deferred effect.
-    ///
-    /// Body IR currently accepts only plain binding patterns for generator clauses. It rejects a whole generator
-    /// expression before evaluating its source when another pattern shape would require a partially represented
-    /// deferred binding protocol; that keeps unsupported forms visible rather than approximating them as a list.
-    fn lower_generator_expr(
-        &mut self,
-        generator: &ast::GeneratorExpr,
-        span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let hir_span_value = hir_span(span);
-        let Some((first_clause, remaining_clauses)) = generator.clauses.split_first() else {
-            return self.unsupported_operand(
-                "generator expression without a for clause".to_string(),
-                scope,
-                hir_span_value,
-                out,
-            );
-        };
-        let ast::ComprehensionClause::For {
-            pattern: first_pattern,
-            iter: first_iter,
-        } = first_clause
-        else {
-            return self.unsupported_operand(
-                "generator expression whose first clause is not a for clause".to_string(),
-                scope,
-                hir_span_value,
-                out,
-            );
-        };
-        let ast::Pattern::Binding(first_name) = &first_pattern.node else {
-            return self.unsupported_operand(
-                "generator for-clause pattern is not a simple binding".to_string(),
-                scope,
-                hir_span_value,
-                out,
-            );
-        };
-        if generator.clauses.iter().any(|clause| {
-            matches!(clause, ast::ComprehensionClause::For { pattern, .. }
-                if !matches!(pattern.node, ast::Pattern::Binding(_)))
-        }) {
-            return self.unsupported_operand(
-                "generator for-clause pattern is not a simple binding".to_string(),
-                scope,
-                hir_span_value,
-                out,
-            );
-        }
-
-        // The first source is the legacy adapter chain's eager boundary. Lowering it before creating the rvalue
-        // preserves source-visible construction timing; all remaining expression lowering below writes only into
-        // `generator_stmts` and therefore happens at poll time.
-        let first_protocol = self.type_info.protocol_iteration(first_iter.span).cloned();
-        let first_is_fallible = first_protocol
-            .as_ref()
-            .is_some_and(|protocol| protocol.fallible_error_type.is_some());
-        let effective_first_iter: &ast::Spanned<ast::Expr> = match (&first_iter.node, first_is_fallible) {
-            (ast::Expr::Try(inner), true) => inner,
-            _ => first_iter,
-        };
-        let source = self.lower_expr_to_operand(effective_first_iter, scope, out);
-
-        let generator_scope = self.new_scope(Some(scope), hir_span_value);
-        let source_local = self.new_temp(
-            self.resolve_ty(effective_first_iter.span),
-            generator_scope,
-            hir_span_value,
-        );
-        self.locals[source_local.index()].origin = bir::LocalOrigin::Captured;
-
-        // Capture every lexical value used after the first source once, at construction. The body cannot read the
-        // enclosing place directly after this point, and restoring the full binding map below prevents generator
-        // clause/capture names from leaking into the following enclosing statement.
-        let enclosing_bindings = self.bindings.clone();
-        let free_names = free_vars_in_generator_deferred_body(generator);
-        let mut captured_operands = Vec::with_capacity(free_names.len());
-        let mut capture_locals = Vec::with_capacity(free_names.len());
-        for name in &free_names {
-            let Some(&outer_local) = self.bindings.get(name) else {
-                // Module/external names remain explicit `External` references when the deferred body is lowered;
-                // there is no local value available to capture and rebind here.
-                continue;
-            };
-            let outer_ty = self.locals[outer_local.index()].ty.clone();
-            let outer_place = bir::Place::from_local(outer_local);
-            let (fact, last_use) = self.ownership_fact_for_place(&outer_place, &outer_ty);
-            captured_operands.push(bir::Operand::place(outer_place, fact, last_use));
-
-            let total_reads = count_reads_in_generator_deferred_body(name, generator);
-            let capture_local =
-                self.declare_new_local_with_reads(name.clone(), outer_ty, generator_scope, hir_span_value, total_reads);
-            self.locals[capture_local.index()].origin = bir::LocalOrigin::Captured;
-            capture_locals.push(capture_local);
-        }
-
-        let first_loop_scope = self.new_scope(Some(generator_scope), hir_span_value);
-        let first_total_reads = count_reads_in_expr(first_name, &generator.expr.node)
-            + count_reads_in_comprehension_clauses(first_name, remaining_clauses);
-        let first_local = self.declare_new_local_with_reads(
-            first_name.clone(),
-            self.resolve_ty(first_pattern.span),
-            first_loop_scope,
-            hir_span(first_pattern.span),
-            first_total_reads,
-        );
-
-        let mut generator_stmts = Vec::new();
-        let iterator_ty = match &first_protocol {
-            Some(protocol) => semantic_type_from_resolved(&protocol.iterator_type),
-            None => self.resolve_ty(effective_first_iter.span),
-        };
-        let iterator_local = self.new_temp(iterator_ty, generator_scope, hir_span_value);
-        match &first_protocol {
-            Some(protocol) => generator_stmts.push(bir::Statement {
-                kind: bir::StatementKind::Call {
-                    destination: Some(bir::Place::from_local(iterator_local)),
-                    callee: bir::Callee::Method(bir::MethodTarget::synthesized(protocol.iter_method.clone())),
-                    args: fixed_elements(vec![bir::Operand::place(
-                        bir::Place::from_local(source_local),
-                        bir::OwnershipFact::Borrow,
-                        false,
-                    )]),
-                    may_panic: false,
-                },
-                span: hir_span_value,
-            }),
-            None => generator_stmts.push(bir::Statement {
-                kind: bir::StatementKind::Assign {
-                    place: bir::Place::from_local(iterator_local),
-                    rvalue: bir::Rvalue::Use(bir::Operand::place(
-                        bir::Place::from_local(source_local),
-                        bir::OwnershipFact::Borrow,
-                        false,
-                    )),
-                },
-                span: hir_span_value,
-            }),
-        }
-
-        self.loop_break_targets.push(None);
-        let mut first_loop_stmts = vec![bir::Statement {
-            kind: bir::StatementKind::IterNext {
-                destination: bir::Place::from_local(first_local),
-                iterator: bir::Operand::place(
-                    bir::Place::from_local(iterator_local),
-                    bir::OwnershipFact::MutBorrow,
-                    false,
-                ),
-                protocol: match &first_protocol {
-                    Some(protocol) => bir::IterProtocol::UserDefined {
-                        next_method: protocol.next_method.clone(),
-                        fallible: first_is_fallible,
-                    },
-                    None => bir::IterProtocol::Builtin,
-                },
-            },
-            span: hir_span_value,
-        }];
-        let terminal = ComprehensionTerminal::GeneratorYield {
-            element: &generator.expr,
-        };
-        self.lower_comprehension_clauses(
-            remaining_clauses,
-            &terminal,
-            first_loop_scope,
-            hir_span_value,
-            &mut first_loop_stmts,
-        );
-        self.insert_scope_drops(&mut first_loop_stmts, first_loop_scope);
-        self.loop_break_targets.pop();
-        generator_stmts.push(bir::Statement {
-            kind: bir::StatementKind::Loop {
-                body: bir::Block {
-                    scope: first_loop_scope,
-                    stmts: first_loop_stmts,
-                },
-            },
-            span: hir_span_value,
-        });
-        self.bindings = enclosing_bindings;
-
-        // `Generator::new` owns a boxed iterator in the legacy runtime, even when every captured source value is
-        // Copy-shaped. Record that allocation fact directly rather than relying on incidental temporary locals.
-        self.record_runtime_requirement(AbiV0RuntimeRequirement::Allocator);
-        let ty = self.resolve_ty(span);
-        self.push_assign_temp(
-            bir::Rvalue::Generator {
-                source,
-                captured_operands,
-                body: Box::new(bir::GeneratorBody {
-                    source_local,
-                    capture_locals,
-                    stmts: generator_stmts,
-                }),
-            },
-            ty,
-            scope,
-            hir_span_value,
-            out,
-        )
-    }
-
-    /// Lower a comprehension/generator clause chain with bindings that are lexical to that expression. The clause
-    /// lowering itself declares each `for` pattern binding through [`Self::declare_new_local_with_reads`] so normal
-    /// operand lowering can resolve it. Those bindings must disappear when the expression ends, however: unlike a
-    /// statement `for`, a comprehension's `x` in `[x for x in values]` cannot shadow an enclosing `x` in the next
-    /// enclosing statement. Preserve the outer lookup map while retaining the locals and ownership facts the nested
-    /// lowering legitimately recorded in the Body IR.
-    fn lower_scoped_comprehension_clauses(
-        &mut self,
-        clauses: &[ast::ComprehensionClause],
-        terminal: &ComprehensionTerminal<'_>,
-        scope: bir::ScopeId,
-        span: HirSourceSpan,
-        out: &mut Vec<bir::Statement>,
-    ) {
-        let enclosing_bindings = self.bindings.clone();
-        self.lower_comprehension_clauses(clauses, terminal, scope, span, out);
-        self.bindings = enclosing_bindings;
-    }
-
-    /// Recursively desugar a comprehension/generator clause chain into nested `Loop`/`If` statements, terminating
-    /// in `terminal`'s compiler-synthesized collection-growth call once every clause has been satisfied for one
-    /// binding combination. `For` clauses reuse [`Self::lower_general_iteration`] (the same builtin-vs-protocol
-    /// iteration primitive [`Self::lower_for`] uses), so comprehensions never duplicate that split. A non-binding
-    /// `For` clause pattern lowers to `Unsupported`, matching [`Self::lower_for`]'s own restriction (destructuring
-    /// patterns need `match`-shaped compilation, out of scope here).
-    fn lower_comprehension_clauses(
-        &mut self,
-        clauses: &[ast::ComprehensionClause],
-        terminal: &ComprehensionTerminal<'_>,
-        scope: bir::ScopeId,
-        span: HirSourceSpan,
-        out: &mut Vec<bir::Statement>,
-    ) {
-        let Some((head, tail)) = clauses.split_first() else {
-            self.lower_comprehension_terminal(terminal, scope, out);
-            return;
-        };
-        match head {
-            ast::ComprehensionClause::If(cond) => {
-                let cond_operand = self.lower_expr_to_operand(cond, scope, out);
-                let then_scope = self.new_scope(Some(scope), span);
-                let mut then_stmts = Vec::new();
-                self.lower_comprehension_clauses(tail, terminal, then_scope, span, &mut then_stmts);
-                out.push(bir::Statement {
-                    kind: bir::StatementKind::If {
-                        cond: cond_operand,
-                        then_block: bir::Block {
-                            scope: then_scope,
-                            stmts: then_stmts,
-                        },
-                        else_block: None,
-                    },
-                    span,
-                });
-            }
-            ast::ComprehensionClause::For { pattern, iter } => {
-                let ast::Pattern::Binding(var_name) = &pattern.node else {
-                    self.push_unsupported_stmt(
-                        "comprehension for-clause pattern is not a simple binding".to_string(),
-                        span,
-                        out,
-                    );
-                    return;
-                };
-                let var_ty = self.resolve_ty(pattern.span);
-                let loop_scope = self.new_scope(Some(scope), span);
-                let total_reads = terminal.count_reads(var_name) + count_reads_in_comprehension_clauses(var_name, tail);
-                let pattern_local =
-                    self.declare_new_local_with_reads(var_name.clone(), var_ty, loop_scope, span, total_reads);
-                self.lower_general_iteration(
-                    iter,
-                    pattern_local,
-                    scope,
-                    loop_scope,
-                    span,
-                    out,
-                    move |builder, loop_scope, body_stmts| {
-                        builder.lower_comprehension_clauses(tail, terminal, loop_scope, span, body_stmts);
-                        builder.insert_scope_drops(body_stmts, loop_scope);
-                    },
-                );
-            }
-        }
-    }
-
-    /// Lower the innermost action of one accepted comprehension/generator binding combination: evaluate the
-    /// element (or key/value) expression(s) and push a compiler-synthesized `push`/`insert`
-    /// [`bir::Callee::Method`] call growing the target collection. The receiver is read as
-    /// [`bir::OwnershipFact::MutBorrow`] since the call mutates the collection in place -- the first real producer
-    /// of that fact in this module (every other place read so far has been `Copy`/`Move`/`Clone`/`Borrow`).
-    fn lower_comprehension_terminal(
-        &mut self,
-        terminal: &ComprehensionTerminal<'_>,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) {
-        match terminal {
-            ComprehensionTerminal::ListPush { list_local, element } => {
-                let element_operand = self.lower_expr_to_operand(element, scope, out);
-                let span = hir_span(element.span);
-                out.push(bir::Statement {
-                    kind: bir::StatementKind::Call {
-                        destination: None,
-                        callee: bir::Callee::Method(bir::MethodTarget::synthesized("push")),
-                        args: fixed_elements(vec![
-                            bir::Operand::place(
-                                bir::Place::from_local(*list_local),
-                                bir::OwnershipFact::MutBorrow,
-                                false,
-                            ),
-                            element_operand,
-                        ]),
-                        may_panic: false,
-                    },
-                    span,
-                });
-            }
-            ComprehensionTerminal::DictInsert { dict_local, key, value } => {
-                let key_operand = self.lower_expr_to_operand(key, scope, out);
-                let value_operand = self.lower_expr_to_operand(value, scope, out);
-                let span = hir_span(value.span);
-                out.push(bir::Statement {
-                    kind: bir::StatementKind::Call {
-                        destination: None,
-                        callee: bir::Callee::Method(bir::MethodTarget::synthesized("insert")),
-                        args: fixed_elements(vec![
-                            bir::Operand::place(
-                                bir::Place::from_local(*dict_local),
-                                bir::OwnershipFact::MutBorrow,
-                                false,
-                            ),
-                            key_operand,
-                            value_operand,
-                        ]),
-                        may_panic: false,
-                    },
-                    span,
-                });
-            }
-            ComprehensionTerminal::GeneratorYield { element } => {
-                let value = self.lower_expr_to_operand(element, scope, out);
-                out.push(bir::Statement {
-                    kind: bir::StatementKind::Yield { value },
-                    span: hir_span(element.span),
-                });
-            }
-        }
-    }
-
-    /// Lower `assert cond[, message]`, recording an [`bir::PanicReason::AssertFailure`] panic fact and a
-    /// [`AbiV0RuntimeRequirement::PanicStrategy`] runtime requirement since every assert can panic. The pattern
-    /// (`assert value is Some(name)`) and `raises` (`assert call() raises E`) forms are not modeled by v0 and lower
-    /// to an explicit unsupported placeholder instead (#1167). The pattern form's placeholder is lossy rather than
-    /// merely incomplete: it discards the names the pattern would bind, so a later read of one lowers against a
-    /// local this body never declared.
-    fn lower_assert(
-        &mut self,
-        assert_stmt: &ast::AssertStmt,
-        scope: bir::ScopeId,
-        span: HirSourceSpan,
-        out: &mut Vec<bir::Statement>,
-    ) {
-        let ast::AssertKind::Condition(cond_expr) = &assert_stmt.kind else {
-            self.push_unsupported_stmt("assert pattern/raises form".to_string(), span, out);
-            return;
-        };
-        let cond = self.lower_expr_to_operand(cond_expr, scope, out);
-        let message = assert_stmt
-            .message
-            .as_ref()
-            .map(|m| self.lower_expr_to_operand(m, scope, out));
-        self.panic_facts.push(bir::PanicFact {
-            span,
-            reason: bir::PanicReason::AssertFailure,
-        });
-        self.record_runtime_requirement(AbiV0RuntimeRequirement::PanicStrategy);
-        out.push(bir::Statement {
-            kind: bir::StatementKind::Assert {
-                cond,
-                message,
-                may_panic: true,
-            },
-            span,
-        });
-    }
-
-    // ---- Callable defaults ----
-
-    /// Lower one source-declared default into a deferred Body-IR computation.
-    ///
-    /// The ordinary function body may not contain this computation: source defaults run only when the matching
-    /// parameter is omitted. While lowering it, callable-local bindings are hidden because the legacy path
-    /// materializes source defaults while assembling call arguments, before the callee frame is bound. A default
-    /// therefore becomes a closed Body-IR computation or a tagged refusal: a callable-local or other external
-    /// source read, every explicitly unsupported Body-IR form, and a default without a usable canonical type fact
-    /// refuse at the default expression's own span. The final condition is deliberately fail-closed: Body IR may
-    /// not make an unchecked source default executable by reconstructing source semantics. This leaves a direct
-    /// consumer no reason to consult AST/HIR/typechecker state or legacy execution.
-    fn lower_callable_default(
-        &mut self,
-        default_expr: Option<&ast::Spanned<ast::Expr>>,
-        scope: bir::ScopeId,
-    ) -> bir::CallableParamDefault {
-        let Some(default_expr) = default_expr else {
-            return bir::CallableParamDefault::Required;
-        };
-
-        let locals_len = self.locals.len();
-        let scopes_len = self.scopes.len();
-        let runtime_requirements_len = self.runtime_requirements.len();
-        let panic_facts_len = self.panic_facts.len();
-        let next_local = self.next_local;
-        let next_scope = self.next_scope;
-        let saved_remaining_reads = self.remaining_reads.clone();
-        let saved_moved_out = self.moved_out.clone();
-        let saved_bindings = std::mem::take(&mut self.bindings);
-        let saved_external_locals = std::mem::take(&mut self.external_locals);
-        let mut stmts = Vec::new();
-        let result = self.lower_expr_to_operand(default_expr, scope, &mut stmts);
-        let mut unresolved_names: Vec<String> = self.external_locals.keys().cloned().collect();
-        unresolved_names.sort();
-        self.bindings = saved_bindings;
-        self.external_locals = saved_external_locals;
-
-        let refusal = first_unsupported_default_statement(&stmts)
-            .or_else(|| {
-                (!unresolved_names.is_empty()).then(|| {
-                    (
-                        hir_span(default_expr.span),
-                        format!(
-                            "default reads Body-IR-external name(s): {}",
-                            unresolved_names.join(", ")
-                        ),
-                    )
-                })
-            })
-            .or_else(|| {
-                self.type_info
-                    .validated_newtype_coercion(default_expr.span)
-                    .is_some()
-                    .then(|| {
-                        (
-                            hir_span(default_expr.span),
-                            "default requires a validated-newtype coercion Body IR does not yet represent".to_string(),
-                        )
-                    })
-            })
-            .or_else(|| {
-                matches!(
-                    self.resolve_ty(default_expr.span),
-                    IncanType::Unknown | IncanType::Never
-                )
-                .then(|| {
-                    (
-                        hir_span(default_expr.span),
-                        "default expression lacks a usable typecheck fact".to_string(),
-                    )
-                })
-            });
-        if let Some((span, description)) = refusal {
-            self.locals.truncate(locals_len);
-            self.scopes.truncate(scopes_len);
-            self.runtime_requirements.truncate(runtime_requirements_len);
-            self.panic_facts.truncate(panic_facts_len);
-            self.next_local = next_local;
-            self.next_scope = next_scope;
-            self.remaining_reads = saved_remaining_reads;
-            self.moved_out = saved_moved_out;
-            return bir::CallableParamDefault::Unsupported { span, description };
-        }
-
-        bir::CallableParamDefault::Source(bir::DefaultComputation {
-            span: hir_span(default_expr.span),
-            stmts,
-            result,
-        })
-    }
-
-    // ---- Expressions ----
-
-    /// Lower one expression into an [`bir::Operand`], dispatching on its AST kind and, where evaluation has side
-    /// effects or must be flattened (calls, binary/unary ops, aggregates), pushing supporting statements into `out`
-    /// first. Expression kinds outside v0's covered subset fall through to [`Self::unsupported_operand`] rather than
-    /// panicking (see this module's module-level docs for the exact covered/uncovered split).
-    fn lower_expr_to_operand(
-        &mut self,
-        expr: &ast::Spanned<ast::Expr>,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let span = hir_span(expr.span);
-        match &expr.node {
-            ast::Expr::Ident(name) => {
-                let place = bir::Place::from_local(self.local_for_name(name, span));
-                let ty = self.resolve_ty(expr.span);
-                let (fact, last_use) = self.ownership_fact_for_place(&place, &ty);
-                bir::Operand::place(place, fact, last_use)
-            }
-            ast::Expr::SelfExpr => {
-                // Resolved exactly like `Ident("self")` — see `BodyBuilder::declare_receiver_local`, which binds
-                // the receiver under the name "self" so this shares `local_for_name`'s ordinary lookup path. A
-                // top-level function body can never actually contain `SelfExpr` (the parser only accepts it inside
-                // a method), so this arm's `local_for_name` fallback to an `External` local is purely defensive.
-                let place = bir::Place::from_local(self.local_for_name("self", span));
-                let ty = self.resolve_ty(expr.span);
-                let (fact, last_use) = self.ownership_fact_for_place(&place, &ty);
-                bir::Operand::place(place, fact, last_use)
-            }
-            ast::Expr::Literal(lit) => match lower_literal(lit) {
-                Some(constant) => bir::Operand::Constant(constant),
-                None => self.unsupported_operand("bytes literal".to_string(), scope, span, out),
-            },
-            ast::Expr::Paren(inner) => self.lower_expr_to_operand(inner, scope, out),
-            ast::Expr::Field(base, name) => {
-                if let Some(target) = self.local_fieldless_enum_variant_target(base, name) {
-                    return self.push_assign_temp(
-                        bir::Rvalue::FieldlessEnumVariant(target),
-                        self.resolve_ty(expr.span),
-                        scope,
-                        span,
-                        out,
-                    );
-                }
-                if let Some(target) = self.local_value_enum_variant_target(base, name) {
-                    return self.push_assign_temp(
-                        bir::Rvalue::ValueEnumVariant(target),
-                        self.resolve_ty(expr.span),
-                        scope,
-                        span,
-                        out,
-                    );
-                }
-                let mut place = self.lower_expr_to_place(base, scope, out);
-                place.projection.push(bir::PlaceElem::Field(name.clone()));
-                let ty = self.resolve_ty(expr.span);
-                let (fact, last_use) = self.ownership_fact_for_place(&place, &ty);
-                bir::Operand::place(place, fact, last_use)
-            }
-            ast::Expr::Index(base, index) => {
-                let index_operand = self.lower_expr_to_operand(index, scope, out);
-                let mut place = self.lower_expr_to_place(base, scope, out);
-                place.projection.push(bir::PlaceElem::Index(Box::new(index_operand)));
-                let ty = self.resolve_ty(expr.span);
-                let (fact, last_use) = self.ownership_fact_for_place(&place, &ty);
-                bir::Operand::place(place, fact, last_use)
-            }
-            ast::Expr::Slice(base, slice) => self.lower_slice(base, slice, expr.span, scope, out),
-            ast::Expr::Unary(op, inner) => {
-                let un_op = lower_unary_op(*op);
-                let operand = self.lower_expr_to_operand(inner, scope, out);
-                let ty = self.resolve_ty(expr.span);
-                self.push_assign_temp(bir::Rvalue::UnaryOp(un_op, operand), ty, scope, span, out)
-            }
-            ast::Expr::Binary(lhs, op, rhs) => self.lower_binary(lhs, *op, rhs, expr.span, scope, out),
-            ast::Expr::Call(callee, type_args, args) => self.lower_call(callee, type_args, args, expr.span, scope, out),
-            ast::Expr::MethodCall(recv, name, type_args, args) => {
-                self.lower_method_call(recv, name, type_args, args, expr.span, scope, out)
-            }
-            ast::Expr::Tuple(items) => self.lower_aggregate(bir::AggregateKind::Tuple, items, expr.span, scope, out),
-            ast::Expr::List(entries) => self.lower_list_literal(entries, expr.span, scope, out),
-            ast::Expr::Dict(entries) => self.lower_dict(entries, expr.span, scope, out),
-            ast::Expr::Set(items) => self.lower_aggregate(bir::AggregateKind::Set, items, expr.span, scope, out),
-            ast::Expr::Constructor(name, args) => self.lower_constructor(name, args, expr.span, scope, out),
-            ast::Expr::ListComp(comp) => self.lower_list_comp(comp, expr.span, scope, out),
-            ast::Expr::DictComp(comp) => self.lower_dict_comp(comp, expr.span, scope, out),
-            ast::Expr::Generator(generator) => self.lower_generator_expr(generator, expr.span, scope, out),
-            ast::Expr::If(if_expr) => self.lower_if_expr(if_expr, scope, expr.span, out),
-            ast::Expr::Loop(loop_expr) => self.lower_loop_expr(loop_expr, scope, expr.span, out),
-            ast::Expr::Try(inner) => self.lower_try(inner, expr.span, scope, out),
-            ast::Expr::FString(parts) => self.lower_fstring(parts, expr.span, scope, out),
-            ast::Expr::Closure(params, body) => self.lower_closure(params, body, expr.span, scope, out),
-            ast::Expr::Partial(partial) => self.lower_partial(partial, expr.span, scope, out),
-            ast::Expr::Match(subject, arms) => self.lower_match(subject, arms, expr.span, scope, out),
-            ast::Expr::Surface(surface) => self.lower_surface_expr(surface, expr.span, scope, out),
-            other => self.unsupported_operand(unsupported_expr_label(other), scope, span, out),
-        }
-    }
-
-    /// Lower an expression that is being used as a place base (the target of `.field`/`[index]` projection or a
-    /// bare name), synthesizing a temporary to hold the value when the expression is not itself place-shaped.
-    fn lower_expr_to_place(
-        &mut self,
-        expr: &ast::Spanned<ast::Expr>,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Place {
-        match &expr.node {
-            ast::Expr::Ident(name) => bir::Place::from_local(self.local_for_name(name, hir_span(expr.span))),
-            ast::Expr::SelfExpr => bir::Place::from_local(self.local_for_name("self", hir_span(expr.span))),
-            ast::Expr::Field(base, name) => {
-                let mut place = self.lower_expr_to_place(base, scope, out);
-                place.projection.push(bir::PlaceElem::Field(name.clone()));
-                place
-            }
-            ast::Expr::Index(base, index) => {
-                let index_operand = self.lower_expr_to_operand(index, scope, out);
-                let mut place = self.lower_expr_to_place(base, scope, out);
-                place.projection.push(bir::PlaceElem::Index(Box::new(index_operand)));
-                place
-            }
-            ast::Expr::Paren(inner) => self.lower_expr_to_place(inner, scope, out),
-            _ => {
-                let ty = self.resolve_ty(expr.span);
-                let operand = self.lower_expr_to_operand(expr, scope, out);
-                self.materialize_operand_to_place(operand, ty, scope, hir_span(expr.span), out)
-            }
-        }
-    }
-
-    /// Ensure `operand` is place-shaped, materializing a fresh temporary holding it first if it is a bare constant.
-    /// Used wherever a value that has already been lowered to an [`bir::Operand`] needs a [`bir::Place`] to project
-    /// further into -- [`Self::lower_expr_to_place`]'s own non-place-shaped fallback, plus tuple-element
-    /// extraction for [`Self::lower_tuple_unpack`]/[`Self::lower_tuple_assign`].
-    fn materialize_operand_to_place(
-        &mut self,
-        operand: bir::Operand,
-        ty: IncanType,
-        scope: bir::ScopeId,
-        span: HirSourceSpan,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Place {
-        match operand {
-            bir::Operand::Place(place_operand) => place_operand.place,
-            constant @ bir::Operand::Constant(_) => {
-                let temp = self.new_temp(ty, scope, span);
-                out.push(bir::Statement {
-                    kind: bir::StatementKind::Assign {
-                        place: bir::Place::from_local(temp),
-                        rvalue: bir::Rvalue::Use(constant),
-                    },
-                    span,
-                });
-                bir::Place::from_local(temp)
-            }
-        }
-    }
-
-    /// Lower a binary-operator expression. Bails out to an explicit unsupported placeholder *before* evaluating
-    /// either operand when `op` has no Body IR v0 handling at all (see [`Self::binary_op_is_supported`]), so an
-    /// unsupported operator's sub-expressions are never partially lowered. Otherwise defers to
-    /// [`Self::lower_binary_from_operands`] for the actual string-helper-or-plain-binop emission, which is also
-    /// shared with [`Self::lower_compound_assignment`].
-    fn lower_binary(
-        &mut self,
-        lhs: &ast::Spanned<ast::Expr>,
-        op: ast::BinaryOp,
-        rhs: &ast::Spanned<ast::Expr>,
-        span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let hir_span_value = hir_span(span);
-        let lhs_ty = self.resolve_ty(lhs.span);
-        let rhs_ty = self.resolve_ty(rhs.span);
-        let result_ty = self.resolve_ty(span);
-
-        // A user-defined operator is a method call, not a primitive operation. The typechecker already resolved
-        // which dunder this spelling dispatches to, so lowering follows that decision rather than falling through
-        // to the primitive operator set -- which would represent `a + b` on two `Vec2` values as machine addition.
-        if let Some(dispatch) = self.type_info.resolved_operator_call(span)
-            && dispatch.kind == ResolvedOperatorKind::Binary
-        {
-            let method = dispatch.method.clone();
-            return self.lower_operator_dispatch(&method, lhs, rhs, result_ty, scope, hir_span_value, out);
-        }
-
-        if !Self::binary_op_is_supported(op, &lhs_ty, &rhs_ty) {
-            return self.unsupported_operand(format!("binary operator {op:?}"), scope, hir_span_value, out);
-        }
-        let lhs_operand = self.lower_expr_to_operand(lhs, scope, out);
-        let rhs_operand = self.lower_expr_to_operand(rhs, scope, out);
-        self.lower_binary_from_operands(
-            op,
-            &lhs_ty,
-            lhs_operand,
-            &rhs_ty,
-            rhs_operand,
-            result_ty,
-            scope,
-            hir_span_value,
-            out,
-        )
-    }
-
-    /// Whether `op` between operands of `lhs_ty`/`rhs_ty` has any Body IR v0 handling (either the string-helper
-    /// path or a direct [`bir::BinOp`] mapping). Checked *before* evaluating operand sub-expressions in both
-    /// [`Self::lower_binary`] and [`Self::lower_compound_assignment`], so an operator v0 does not model never
-    /// causes its operands' side effects (calls, reads) to be lowered on the way to an unsupported placeholder.
-    /// Lower a user-defined operator to the dunder method call the typechecker resolved for it.
-    ///
-    /// RFC 028 lets a type define `__add__`, `__and__`, `__contains__` and friends, and the typechecker records
-    /// which method one operator spelling dispatches to. Body IR must follow that decision: representing `a + b` on
-    /// two `Vec2` values as [`bir::BinOp::Add`] would claim a primitive machine operation where the source calls a
-    /// method, which is a wrong representation rather than an honest refusal — no `Unsupported` marker, nothing for
-    /// a consumer to notice.
-    ///
-    /// The left operand becomes the receiver and the right becomes the single argument, matching how
-    /// [`Self::lower_method_call`] arranges an ordinary method call: `args[0]` is the receiver, borrowed. The
-    /// binding is [`bir::ArgumentBinding::UnresolvedPositional`] because an operator spelling names no parameter
-    /// and this stage resolves no declared slot for it.
-    #[allow(clippy::too_many_arguments)]
-    fn lower_operator_dispatch(
-        &mut self,
-        method: &str,
-        lhs: &ast::Spanned<ast::Expr>,
-        rhs: &ast::Spanned<ast::Expr>,
-        result_ty: IncanType,
-        scope: bir::ScopeId,
-        span: HirSourceSpan,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        // Source evaluation observes the receiver before the argument, exactly as for a written method call.
-        let receiver_place = self.lower_expr_to_place(lhs, scope, out);
-        let receiver = bir::Operand::place(receiver_place, bir::OwnershipFact::Borrow, false);
-        let argument = self.lower_expr_to_operand(rhs, scope, out);
-        self.push_call_temp(
-            bir::Callee::Method(bir::MethodTarget::synthesized(method)),
-            vec![bir::ArgumentElement::One(receiver), bir::ArgumentElement::One(argument)],
-            result_ty,
-            scope,
-            span,
-            false,
-            out,
-        )
-    }
-
-    fn binary_op_is_supported(op: ast::BinaryOp, lhs_ty: &IncanType, rhs_ty: &IncanType) -> bool {
-        (is_string_like(lhs_ty) && is_string_like(rhs_ty) && string_helper_for_binop(op).is_some())
-            || lower_binary_op(op).is_some()
-    }
-
-    /// Emit the result of a binary operator given already-lowered operands: an explicit [`bir::Callee::Helper`]
-    /// call (with runtime requirements recorded) when both operand types are string-like and `op` has a
-    /// compiler-owned string helper (see [`string_helper_for_binop`]) -- Body IR's compiler-owned-runtime-operation
-    /// requirement (#653 criterion 3) applied to string operators specifically -- otherwise a plain
-    /// [`bir::Rvalue::BinaryOp`], with a division/modulo panic fact recorded when [`bir::BinOp::may_panic`] holds.
-    /// Callers are expected to have already checked [`Self::binary_op_is_supported`]; an operator with neither
-    /// handling still falls back to an explicit unsupported placeholder defensively rather than panicking.
-    #[allow(clippy::too_many_arguments)]
-    fn lower_binary_from_operands(
-        &mut self,
-        op: ast::BinaryOp,
-        lhs_ty: &IncanType,
-        lhs_operand: bir::Operand,
-        rhs_ty: &IncanType,
-        rhs_operand: bir::Operand,
-        result_ty: IncanType,
-        scope: bir::ScopeId,
-        span: HirSourceSpan,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        if is_string_like(lhs_ty)
-            && is_string_like(rhs_ty)
-            && let Some(helper) = string_helper_for_binop(op)
-        {
-            self.record_runtime_requirement(AbiV0RuntimeRequirement::RuntimeHelper(helper.as_str().to_string()));
-            self.record_runtime_requirement(AbiV0RuntimeRequirement::Allocator);
-            return self.push_call_temp(
-                bir::Callee::Helper(helper),
-                fixed_elements(vec![lhs_operand, rhs_operand]),
-                result_ty,
-                scope,
-                span,
-                false,
-                out,
-            );
-        }
-
-        let Some(bin_op) = lower_binary_op(op) else {
-            return self.unsupported_operand(format!("binary operator {op:?}"), scope, span, out);
-        };
-        if bin_op.may_panic() {
-            self.panic_facts.push(bir::PanicFact {
-                span,
-                reason: bir::PanicReason::DivisionOrModulo,
-            });
-            self.record_runtime_requirement(AbiV0RuntimeRequirement::PanicStrategy);
-        }
-        self.push_assign_temp(
-            bir::Rvalue::BinaryOp(bin_op, lhs_operand, rhs_operand),
-            result_ty,
-            scope,
-            span,
-            out,
-        )
-    }
-
-    /// Lower planned call arguments in written source order, then place them into declaration-slot order.
-    ///
-    /// Both orders are part of the source contract and they differ whenever a caller writes named arguments out of
-    /// declaration order. Argument expressions are therefore lowered here strictly left to right, so the emitted
-    /// statement sequence observes written evaluation order, while the returned operand vector is in declaration
-    /// order and the returned [`bir::ArgumentBinding`] records which slot each operand fills and where it was
-    /// written. A declaration slot the call site never supplied becomes a defaulted slot rather than an operand:
-    /// this call site evaluates nothing for it, so it has no ownership fact to record and the default's computation
-    /// stays owned by the declaration.
-    ///
-    /// Because ownership is decided during that written-order pass, each operand's [`bir::OwnershipFact`] and
-    /// last-use marker are sequenced by `written_position` and **not** by operand index -- see
-    /// [`bir::ArgumentBinding`]'s own docs, which state the invariant a consumer has to honor.
-    fn lower_planned_args(
-        &mut self,
-        planned: &[(usize, &ast::Spanned<ast::Expr>)],
-        slot_count: usize,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> Result<(Vec<bir::Operand>, bir::ArgumentBinding), String> {
-        // Both planners derive their slots from the same declaration surface they report the count of, so an
-        // out-of-range slot is unreachable. It is still refused rather than skipped: dropping the operand while
-        // leaving the statements that computed it in `out` would produce a silently wrong call, which is the worst
-        // failure mode this node has.
-        if let Some((slot, _)) = planned.iter().find(|(slot, _)| *slot >= slot_count) {
-            return Err(format!(
-                "argument bound to declaration slot {slot} outside the callee's {slot_count} declared slots"
-            ));
-        }
-        let mut lowered: Vec<Option<(bir::Operand, usize)>> = (0..slot_count).map(|_| None).collect();
-        for (written_position, (slot, expr)) in planned.iter().enumerate() {
-            let operand = self.lower_expr_to_operand(expr, scope, out);
-            if let Some(entry) = lowered.get_mut(*slot) {
-                *entry = Some((operand, written_position));
-            }
-        }
-
-        let mut operands = Vec::with_capacity(planned.len());
-        let mut arguments = Vec::with_capacity(planned.len());
-        let mut defaulted_slots = Vec::new();
-        for (slot, entry) in lowered.into_iter().enumerate() {
-            match entry {
-                Some((operand, written_position)) => {
-                    operands.push(operand);
-                    arguments.push(bir::BoundArgument { slot, written_position });
-                }
-                None => defaulted_slots.push(slot),
-            }
-        }
-        Ok((
-            operands,
-            bir::ArgumentBinding::Resolved {
-                arguments,
-                defaulted_slots,
-            },
-        ))
-    }
-
-    /// Build the RFC 120 identity of a declaration *this* module owns, from the declaration the call selected.
-    ///
-    /// The span is the one the caller already resolved to build [`DirectCallDeclaration::direct_call_id`], so the two
-    /// facts on a `NamedCallableTarget` are derived from one decision and cannot name different declarations. That
-    /// matters because locality must not be inferred from the recorded source target: an import binding of the same
-    /// spelling wins in `TypeChecker::source_target_for_symbol` regardless of what the call actually bound, so a local
-    /// declaration shadowed by a same-name import would otherwise be given the *import's* identity.
-    fn local_callable_identity(&self, declared_name: &str, declaration_span: ast::Span) -> CanonicalSymbolId {
-        CanonicalSymbolId::module_declaration(
-            self.module_path.to_vec(),
-            declared_name,
-            SemanticSourceTargetKind::Function,
-            HirSourceSpan::new(declaration_span.start, declaration_span.end),
-        )
-    }
-
-    /// Return the RFC 120 identity of an imported callable, when import resolution proved one.
-    ///
-    /// Only reached when this module declares no function of that spelling, so there is no local declaration for an
-    /// import to be confused with. The proven identity carries its own declaration kind, so a binding that resolved to
-    /// something other than a function yields nothing rather than a function identity.
-    ///
-    /// An overloaded import is refused upstream, in `TypeChecker::dependency_member_identity`: an overloaded symbol
-    /// keeps only the first candidate's span, so an identity minted from it would name an arbitrary overload.
-    fn imported_callable_identity(&self, call_site_name: &str) -> Option<CanonicalSymbolId> {
-        let identity = self.type_info.resolved_import_identity(call_site_name)?;
-        (identity.kind == SemanticSourceTargetKind::Function).then(|| identity.clone())
-    }
-
-    /// Resolve the declaration surface and exact local identity for a direct named call.
-    ///
-    /// A direct executable target must be physically represented by this Body-IR module. Imports and unresolved
-    /// names deliberately retain their existing call representation with no direct declaration identity, so this
-    /// frontend does not turn a source-representation gap into a new source diagnostic. The replacement executor
-    /// then refuses those targets at the original call span; only compiler-recognized `range` has a separate
-    /// explicit Body-IR builtin target fact.
-    ///
-    /// Overloads are why this is resolved per call site rather than per name. `function_bindings` is keyed by bare
-    /// source name, so for two same-name declarations it holds only one of them; binding a call against the wrong
-    /// overload's parameter *names* would silently reorder its arguments, turning an honest refusal into a wrong
-    /// answer. The typechecker already records which overload it selected for this call span, so this follows that
-    /// decision to the declaration and reads that declaration's own signature. If a name is overloaded but no
-    /// selection was recorded, this fails closed rather than picking one.
-    fn declared_slots_for_direct_call(&self, name: &str, span: ast::Span) -> Result<DirectCallDeclaration, String> {
-        let declarations = &self.type_info.declarations;
-        let local_declarations = self.local_function_declarations.get(name);
-        let Some(local_declarations) = local_declarations else {
-            return Ok(DirectCallDeclaration {
-                slots: declarations
-                    .function_bindings
-                    .get(name)
-                    .map(|binding| binding.params.iter().map(DeclaredSlot::from_checked_param).collect()),
-                direct_call_id: None,
-                builtin: (name == "range"
-                    && self.type_info.source_target(span).is_none()
-                    && !declarations.function_bindings.contains_key(name)
-                    && !declarations.function_overloads.contains_key(name))
-                .then_some(bir::NamedCallableBuiltin::Range),
-                canonical: self.imported_callable_identity(name),
-            });
-        };
-        let is_overloaded = local_declarations.len() > 1;
-
-        if is_overloaded {
-            let Some(selected) = self.type_info.selected_function_emitted_name(span) else {
-                return Err(format!(
-                    "call to overloaded function `{name}` whose selected declaration was not resolved"
-                ));
-            };
-            let selected_span = local_declarations.iter().find(|candidate_span| {
-                declarations
-                    .function_emitted_names
-                    .get(&(candidate_span.start, candidate_span.end))
-                    .is_some_and(|emitted| emitted == selected)
-            });
-            let Some(selected_span) = selected_span else {
-                return Err(format!(
-                    "call to overloaded function `{name}` whose selected declaration could not be located"
-                ));
-            };
-            let Some(binding) = declarations
-                .function_bindings_by_span
-                .get(&(selected_span.start, selected_span.end))
-            else {
-                return Err(format!(
-                    "call to overloaded function `{name}` whose selected declaration has no checked signature"
-                ));
-            };
-            return Ok(DirectCallDeclaration {
-                slots: Some(binding.params.iter().map(DeclaredSlot::from_checked_param).collect()),
-                direct_call_id: Some(CompilerNodeId::declaration_span(
-                    self.module_identity,
-                    selected_span.start,
-                    selected_span.end,
-                )),
-                builtin: None,
-                // The identity anchors to the declaration span, so the selected overload is as nameable as any other
-                // declaration; it is the *spelling* that cannot separate them, and the spelling is not the identity.
-                canonical: Some(self.local_callable_identity(name, *selected_span)),
-            });
-        }
-
-        let [declaration_span] = local_declarations.as_slice() else {
-            return Err(format!(
-                "direct call to `{name}` has no unambiguous same-module declaration identity"
-            ));
-        };
-        let Some(binding) = declarations
-            .function_bindings_by_span
-            .get(&(declaration_span.start, declaration_span.end))
-        else {
-            return Err(format!(
-                "same-module declaration `{name}` has no checked callable signature"
-            ));
-        };
-        Ok(DirectCallDeclaration {
-            slots: Some(binding.params.iter().map(DeclaredSlot::from_checked_param).collect()),
-            direct_call_id: Some(CompilerNodeId::declaration_span(
-                self.module_identity,
-                declaration_span.start,
-                declaration_span.end,
-            )),
-            builtin: None,
-            canonical: Some(self.local_callable_identity(name, *declaration_span)),
-        })
-    }
-
-    /// Bind a call's arguments against a declared parameter surface, falling back to positional lowering when there
-    /// is none to bind against.
-    ///
-    /// Shared by the direct-call and method paths so both treat an unresolved or rest-bearing signature the same
-    /// way. A rest (`*args`/`**kwargs`) parameter means a written argument no longer corresponds one-to-one with a
-    /// declared slot, so those calls keep lowering their arguments — refusing them would drop a delivered language
-    /// capability — but record [`bir::ArgumentBinding::UnresolvedPositional`] rather than a slot map this stage did
-    /// not compute. Spread arguments lower there, because a spread genuinely has no slot to bind to. A *named*
-    /// argument with no spread beside it is still refused: its arity is perfectly well known, so binding it into a
-    /// rest parameter is variadic-binding work this issue does not own.
-    fn bind_declared_args(
-        &mut self,
-        callee: &str,
-        declared: Option<Vec<DeclaredSlot>>,
-        args: &[ast::CallArg],
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> Result<(Vec<bir::ArgumentElement>, bir::ArgumentBinding), String> {
-        let has_rest = declared
-            .as_ref()
-            .is_some_and(|slots| slots.iter().any(|slot| slot.is_rest));
-        let fixed_slots = declared.filter(|_| !has_rest);
-        let Some(slots) = fixed_slots else {
-            let elements = self.lower_spread_capable_args(args, scope, out);
-            return Ok((elements, bir::ArgumentBinding::UnresolvedPositional));
-        };
-        // A spread whose shape the typechecker proved is an ordinary fixed-arity call in disguise: `add(*(1, 2))`
-        // really is `add(1, 2)`. Expanding it here means it binds through the same declaration-slot planner as any
-        // other call, instead of being pushed onto the runtime-arity path it does not belong on.
-        let expanded: Vec<ast::CallArg> = args
-            .iter()
-            .flat_map(|arg| match expand_shaped_spread(self.type_info, arg) {
-                Some(expansion) => expansion,
-                None => vec![arg.clone()],
-            })
-            .collect();
-        let planned = plan_declared_args(callee, &slots, &expanded)?;
-        let (operands, binding) = self
-            .lower_planned_args(&planned, slots.len(), scope, out)
-            .map_err(|description| format!("{callee}: {description}"))?;
-        Ok((fixed_elements(operands), binding))
-    }
-
-    /// Resolve a call site's explicit type arguments to semantic types, or describe why they cannot be represented.
-    ///
-    /// Explicit type arguments are part of a call's resolved identity, so Body IR takes the typechecker's
-    /// monomorphized selection rather than re-lowering the written AST type nodes -- which is also the only way a
-    /// `_` placeholder resolves to a real type instead of an unknown. A call that wrote type arguments the
-    /// typechecker did not resolve is refused by name rather than represented with a guess.
-    fn call_site_type_arguments(
-        &self,
-        span: ast::Span,
-        type_args: &[ast::Spanned<ast::Type>],
-    ) -> Result<Vec<IncanType>, String> {
-        if type_args.is_empty() {
-            return Ok(Vec::new());
-        }
-        let Some(resolved) = self
-            .type_info
-            .calls
-            .call_site_monomorph_type_args
-            .get(&(span.start, span.end))
-        else {
-            return Err("call with unresolved explicit type arguments".to_string());
-        };
-        Ok(resolved.iter().map(semantic_type_from_resolved).collect())
-    }
-
-    /// Lower a `model`/`class` construction into a [`bir::AggregateKind::Constructor`] aggregate.
-    ///
-    /// Source-level construction is named-only, so the argument-to-field binding is the whole representation
-    /// problem. Lowering consumes the typechecker's own recorded decision
-    /// ([`TypeCheckInfo::constructor_field_binding`](crate::frontend::typechecker::TypeCheckInfo::constructor_field_binding))
-    /// rather than re-resolving field aliases or rediscovering declared field order, both of which live in the
-    /// symbol table this stage deliberately cannot reach. Operands are emitted in declared field order while the
-    /// argument expressions are lowered in written source order, exactly as for a call.
-    fn lower_nominal_construction(
-        &mut self,
-        name: &str,
-        args: &[ast::CallArg],
-        span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let hir_span_value = hir_span(span);
-        let Some(field_binding) = self.type_info.constructor_field_binding(span).cloned() else {
-            return self.unsupported_operand(
-                format!("construction of `{name}` with an unresolved field layout"),
-                scope,
-                hir_span_value,
-                out,
-            );
-        };
-
-        // The typechecker records one slot per *written* argument, so a spread -- which supplies an unknown number
-        // of fields -- can never appear in a recorded binding. Refuse it by name; #1159 owns spread representation.
-        let mut written_exprs = Vec::with_capacity(args.len());
-        for arg in args {
-            match arg {
-                ast::CallArg::Positional(expr) | ast::CallArg::Named(_, expr) => written_exprs.push(expr),
-                ast::CallArg::PositionalUnpack(_) => {
-                    return self.unsupported_operand(
-                        format!("construction of `{name}` with a positional argument spread"),
-                        scope,
-                        hir_span_value,
-                        out,
-                    );
-                }
-                ast::CallArg::KeywordUnpack(_) => {
-                    return self.unsupported_operand(
-                        format!("construction of `{name}` with a keyword argument spread"),
-                        scope,
-                        hir_span_value,
-                        out,
-                    );
-                }
-            }
-        }
-        if written_exprs.len() != field_binding.argument_slots.len() {
-            return self.unsupported_operand(
-                format!("construction of `{name}` with an unresolved field layout"),
-                scope,
-                hir_span_value,
-                out,
-            );
-        }
-
-        let planned: Vec<(usize, &ast::Spanned<ast::Expr>)> = field_binding
-            .argument_slots
-            .iter()
-            .copied()
-            .zip(written_exprs)
-            .collect();
-        let (operands, binding) = match self.lower_planned_args(&planned, field_binding.field_count, scope, out) {
-            Ok(bound) => bound,
-            Err(description) => {
-                return self.unsupported_operand(
-                    format!("construction of `{name}`: {description}"),
-                    scope,
-                    hir_span_value,
-                    out,
-                );
-            }
-        };
-        let ty = self.resolve_ty(span);
-        // A constructor field binding proves argument slots, but not that this constructor names one of the plain
-        // source-local models this Body-IR module retained. Preserve an identity only from that local registry;
-        // imports, aliases, classes, generic models, and absent/malformed names remain represented with `None` so
-        // a direct executor can refuse at this construction span rather than guessing from `name`.
-        let direct_declaration_id = self.local_nominal_declarations.get(name).and_then(|declaration| {
-            (declaration.fields.len() == field_binding.field_count).then(|| declaration.direct_declaration_id.clone())
-        });
-        self.push_assign_temp(
-            bir::Rvalue::Aggregate(
-                bir::AggregateKind::Constructor(bir::ConstructorTarget {
-                    name: name.to_string(),
-                    direct_declaration_id,
-                    binding,
-                }),
-                fixed_elements(operands),
-            ),
-            ty,
-            scope,
-            hir_span_value,
-            out,
-        )
-    }
-
-    /// Lower call arguments positionally, admitting spreads, for a call whose arity is not statically known.
-    ///
-    /// Every argument keeps its written form: a positional value, a named value, or a spread. None of them can be
-    /// resolved to a declared slot here, because a spread supplies an unknown number of arguments at runtime —
-    /// which is exactly why the resulting call records [`bir::ArgumentBinding::UnresolvedPositional`] rather than a
-    /// slot map asserting a binding nobody checked. A name is preserved on its element rather than discarded, so a
-    /// later consumer can still bind it once the arity is known.
-    fn lower_spread_capable_args(
-        &mut self,
-        args: &[ast::CallArg],
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> Vec<bir::ArgumentElement> {
-        let mut elements = Vec::with_capacity(args.len());
-        for arg in args {
-            match arg {
-                ast::CallArg::Positional(expr) => {
-                    elements.push(bir::ArgumentElement::One(self.lower_expr_to_operand(expr, scope, out)));
-                }
-                ast::CallArg::PositionalUnpack(source) => {
-                    elements.push(self.lower_spread_element(source, bir::SpreadKind::Sequence, scope, out));
-                }
-                ast::CallArg::KeywordUnpack(source) => {
-                    elements.push(self.lower_spread_element(source, bir::SpreadKind::Mapping, scope, out));
-                }
-                ast::CallArg::Named(name, expr) => {
-                    let operand = self.lower_expr_to_operand(expr, scope, out);
-                    elements.push(bir::ArgumentElement::Named {
-                        name: name.clone(),
-                        operand,
-                    });
-                }
-            }
-        }
-        elements
-    }
-
-    /// Return the exact retained target for a qualified local fieldless normal-enum member, if safe to materialize.
-    ///
-    /// A bare type-name receiver and source-local registry membership are both required. This leaves ordinary forms
-    /// not represented by the registry as generic field accesses that the direct executor visibly refuses, while
-    /// preserving exact declaration identities for the one bounded unit-variant carrier profile.
-    fn local_fieldless_enum_variant_target(
-        &self,
-        base: &ast::Spanned<ast::Expr>,
-        variant_name: &str,
-    ) -> Option<bir::FieldlessEnumVariantTarget> {
-        let ast::Expr::Ident(enum_name) = &base.node else {
-            return None;
-        };
-        if self.bindings.contains_key(enum_name)
-            || !matches!(self.type_info.ident_kind(base.span), Some(IdentKind::TypeName))
-        {
-            return None;
-        }
-        let declaration = self.local_fieldless_enum_declarations.get(enum_name)?;
-        let variant = declaration
-            .variants
-            .iter()
-            .find(|variant| variant.name == variant_name)?;
-        Some(bir::FieldlessEnumVariantTarget {
-            enum_declaration_id: declaration.direct_declaration_id.clone(),
-            variant_declaration_id: variant.direct_declaration_id.clone(),
-            enum_name: declaration.name.clone(),
-            variant_name: variant.name.clone(),
-        })
-    }
-
-    /// Return the exact retained target for a qualified local RFC 032 value-enum member, if this spelling is safe to
-    /// materialize directly.
-    ///
-    /// The source-local registry is deliberately the only lookup used here. A function-local binding wins over a
-    /// same-spelling declaration, and any import, alias, ordinary enum, payload member, or behavior-bearing enum is
-    /// absent from the registry. The resulting rvalue stores both declaration identities for runtime revalidation;
-    /// it does not make the spelling itself an execution authority.
-    fn local_value_enum_variant_target(
-        &self,
-        base: &ast::Spanned<ast::Expr>,
-        variant_name: &str,
-    ) -> Option<bir::ValueEnumVariantTarget> {
-        let ast::Expr::Ident(enum_name) = &base.node else {
-            return None;
-        };
-        if self.bindings.contains_key(enum_name) {
-            return None;
-        }
-        if !matches!(self.type_info.ident_kind(base.span), Some(IdentKind::TypeName)) {
-            return None;
-        }
-        let declaration = self.local_value_enum_declarations.get(enum_name)?;
-        let variant = declaration
-            .variants
-            .iter()
-            .find(|variant| variant.name == variant_name)?;
-        Some(bir::ValueEnumVariantTarget {
-            enum_declaration_id: declaration.direct_declaration_id.clone(),
-            variant_declaration_id: variant.direct_declaration_id.clone(),
-            enum_name: declaration.name.clone(),
-            variant_name: variant.name.clone(),
-        })
-    }
-
-    /// Lower a call to a locally held callable value, a nominal construction, or a direct named function.
-    ///
-    /// A bare identifier that resolves to one of this body's locals is deliberately a
-    /// [`bir::CallableTarget::Local`] call: it carries the local read's ownership fact, so a closure's lexical
-    /// environment is not lost by pretending the identifier were a declaration. Its callable signature also
-    /// enforces the stored value's fixed callable contract before any call arguments are lowered. An identifier the
-    /// typechecker resolved to a `model`/`class` construction lowers to a constructor aggregate instead of a call
-    /// (see [`Self::lower_nominal_construction`]) -- construction is not invocation, and representing it as a call
-    /// would invite a consumer to execute it as one. Any other bare identifier remains a direct
-    /// [`bir::Callee::Function`] call.
-    ///
-    /// Every one of those paths binds its arguments through the same [`plan_declared_args`] planner and records the
-    /// result as a [`bir::ArgumentBinding`], so named, out-of-order, and defaulted spellings resolve identically
-    /// regardless of how the callee was reached. A direct call whose signature the typechecker did not resolve
-    /// (notably a builtin) still lowers its arguments faithfully, recording
-    /// [`bir::ArgumentBinding::UnresolvedPositional`], and refuses only a named spelling it cannot bind without one.
-    /// Argument spreads lower as [`bir::ArgumentElement::Spread`] elements, since a spread has no declared slot to
-    /// bind to by construction. A non-identifier callee remains an explicit unsupported form; v0 has no
-    /// dynamic-call-target node for it yet.
-    fn lower_call(
-        &mut self,
-        callee: &ast::Spanned<ast::Expr>,
-        type_args: &[ast::Spanned<ast::Type>],
-        args: &[ast::CallArg],
-        span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let hir_span_value = hir_span(span);
-        let ast::Expr::Ident(name) = &callee.node else {
-            return self.unsupported_operand("indirect call target".to_string(), scope, hir_span_value, out);
-        };
-        let name = name.clone();
-
-        // A recorded constructor field binding is the typechecker's own statement that this spelling constructs a
-        // nominal value, which is what distinguishes `P(x=1)` from a call to a function that happens to be named
-        // `P`. A construction may carry call-site type arguments (`Box[int]()` is accepted), but the typechecker
-        // records no monomorphization for them, and the constructed value's own type already carries the resolved
-        // arguments -- so this deliberately does not duplicate them on the constructor target rather than claiming
-        // construction cannot be generic.
-        if self.type_info.constructor_field_binding(span).is_some() {
-            return self.lower_nominal_construction(&name, args, span, scope, out);
-        }
-
-        // `Ok` and `Err` are intrinsic Result constructors, not ordinary direct calls. Retain that checked
-        // distinction explicitly: a same-spelled source binding (a local callable, local function, or imported
-        // target) must remain on the normal call path and refuse unless its own direct callable facts are available.
-        // The direct runtime never resolves a constructor name dynamically.
-        if !self.bindings.contains_key(&name)
-            && !self.local_function_declarations.contains_key(&name)
-            && self.type_info.source_target(span).is_none()
-            && type_args.is_empty()
-            && let Some(kind) = result_variant_kind(&name)
-        {
-            let result_ty = self.resolve_ty(span);
-            let Some((ok_type, error_type)) = result_type_parts(&result_ty) else {
-                return self.unsupported_operand(
-                    format!("intrinsic Result constructor `{name}` without a resolved Result carrier"),
-                    scope,
-                    hir_span_value,
-                    out,
-                );
-            };
-            let [ast::CallArg::Positional(payload)] = args else {
-                return self.unsupported_operand(
-                    format!("intrinsic Result constructor `{name}` requires one positional payload"),
-                    scope,
-                    hir_span_value,
-                    out,
-                );
-            };
-            let payload = self.lower_expr_to_operand(payload, scope, out);
-            return self.push_assign_temp(
-                bir::Rvalue::ResultVariant(bir::ResultVariant {
-                    kind,
-                    payload,
-                    ok_type: ok_type.clone(),
-                    error_type: error_type.clone(),
-                }),
-                result_ty,
-                scope,
-                hir_span_value,
-                out,
-            );
-        }
-
-        let resolved_type_args = match self.call_site_type_arguments(span, type_args) {
-            Ok(resolved_type_args) => resolved_type_args,
-            Err(description) => {
-                return self.unsupported_operand(description, scope, hir_span_value, out);
-            }
-        };
-
-        if let Some(&local) = self.bindings.get(&name) {
-            let local_ty = self.locals[local.index()].ty.clone();
-            let IncanType::Function { params, return_type: _ } = local_ty else {
-                return self.unsupported_operand(
-                    format!("call to non-callable local `{name}`"),
-                    scope,
-                    hir_span_value,
-                    out,
-                );
-            };
-            let slots: Vec<DeclaredSlot> = params.iter().map(DeclaredSlot::from_semantic_param).collect();
-            let planned = match plan_declared_args(&format!("local callable `{name}`"), &slots, args) {
-                Ok(planned) => planned,
-                Err(description) => {
-                    return self.unsupported_operand(description, scope, hir_span_value, out);
-                }
-            };
-
-            // Source evaluation observes the callable value before its arguments. The target read also performs the
-            // one ownership/last-use decision for that lexical environment, which `CallableTarget::Local` preserves
-            // for a later executor instead of re-deriving it from the local's source spelling.
-            let place = bir::Place::from_local(local);
-            let (fact, last_use) = self.ownership_fact_for_place(&place, &self.locals[local.index()].ty.clone());
-            let (operands, binding) = match self.lower_planned_args(&planned, slots.len(), scope, out) {
-                Ok(bound) => bound,
-                Err(description) => {
-                    return self.unsupported_operand(description, scope, hir_span_value, out);
-                }
-            };
-            let callee = bir::Callee::Function(bir::CallableTarget::Local(bir::LocalCallableTarget {
-                operand: bir::PlaceOperand { place, fact, last_use },
-                binding,
-            }));
-            let ty = self.resolve_ty(span);
-            return self.push_call_temp(callee, fixed_elements(operands), ty, scope, hir_span_value, false, out);
-        }
-
-        // A name that resolves to a nominal type but has no recorded field binding is a construction the checker
-        // declined to bind (a duplicate or unknown field). Refusing it as a call to an unknown function would name
-        // the wrong construct entirely.
-        if self.type_info.declarations.class_layouts.contains_key(&name)
-            || self.type_info.declarations.model_field_visibilities.contains_key(&name)
-        {
-            return self.unsupported_operand(
-                format!("construction of `{name}` with an unresolved field layout"),
-                scope,
-                hir_span_value,
-                out,
-            );
-        }
-
-        let declaration = match self.declared_slots_for_direct_call(&name, span) {
-            Ok(declaration) => declaration,
-            Err(description) => {
-                return self.unsupported_operand(description, scope, hir_span_value, out);
-            }
-        };
-        let (operands, binding) =
-            match self.bind_declared_args(&format!("function `{name}`"), declaration.slots, args, scope, out) {
-                Ok(bound) => bound,
-                Err(description) => {
-                    return self.unsupported_operand(description, scope, hir_span_value, out);
-                }
-            };
-
-        let ty = self.resolve_ty(span);
-        self.push_call_temp(
-            bir::Callee::Function(bir::CallableTarget::Named(bir::NamedCallableTarget {
-                name,
-                direct_call_id: declaration.direct_call_id,
-                canonical: declaration.canonical,
-                builtin: declaration.builtin,
-                type_args: resolved_type_args,
-                binding,
-            })),
-            operands,
-            ty,
-            scope,
-            hir_span_value,
-            false,
-            out,
-        )
-    }
-
-    /// Return the typechecker's callable type for a closure or local partial value that Body IR constructs itself.
-    ///
-    /// Local partials use the typechecker's canonical full signature with overrideable preset-default slots, so the
-    /// binding, its [`bir::Rvalue::Closure`], and a later [`Self::lower_call`] share one arity/default contract.
-    fn callable_value_ty(&self, expr: &ast::Spanned<ast::Expr>) -> Option<IncanType> {
-        match &expr.node {
-            ast::Expr::Closure(_, _) | ast::Expr::Partial(_) => Some(self.resolve_ty(expr.span)),
-            _ => None,
-        }
-    }
-
-    /// Lower a method call `recv.name(args)` to a [`bir::Callee::Method`] call, with the receiver prepended to
-    /// `args[0]` as a [`bir::OwnershipFact::Borrow`] operand (see the inline comment on the receiver-borrow decision
-    /// below).
-    ///
-    /// Argument binding goes through the same [`plan_declared_args`] planner every other call shape uses, against
-    /// the typechecker's own rest-aware call-site signature for this span -- which already has the receiver's
-    /// generic arguments substituted, so a generic method's slots are concrete here. The receiver is deliberately
-    /// outside the recorded binding: its slots index the method's declared parameters, so a consumer reads
-    /// `args[0]` as the receiver and `args[1..]` as the bound arguments. A method call whose signature the
-    /// typechecker did not record still lowers positional arguments faithfully and refuses only the spellings it
-    /// cannot bind — a named spelling with no spread beside it — matching [`Self::lower_call`]'s treatment of an
-    /// unresolved direct callee. Spread arguments lower here too, after the receiver.
-    #[allow(clippy::too_many_arguments)]
-    fn lower_method_call(
-        &mut self,
-        recv: &ast::Spanned<ast::Expr>,
-        name: &str,
-        type_args: &[ast::Spanned<ast::Type>],
-        args: &[ast::CallArg],
-        span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let hir_span_value = hir_span(span);
-        let resolved_type_args = match self.call_site_type_arguments(span, type_args) {
-            Ok(resolved_type_args) => resolved_type_args,
-            Err(_) => {
-                return self.unsupported_operand(
-                    "method call with unresolved explicit type arguments".to_string(),
-                    scope,
-                    hir_span_value,
-                    out,
-                );
-            }
-        };
-
-        let declared: Option<Vec<DeclaredSlot>> = self
-            .type_info
-            .call_site_callable_params(span)
-            .map(|params| params.iter().map(DeclaredSlot::from_checked_param).collect());
-
-        // The receiver is read before the arguments, matching source evaluation order: `recv.m(f())` observes the
-        // receiver place first. Method receivers are treated as borrowed rather than moved/cloned, mirroring how the
-        // existing Rust-emission backend's ownership planner treats most method receivers
-        // (`src/backend/ir/ownership.rs`) -- see this module's rustdoc for the full precedent discussion.
-        let receiver_operand = if let ast::Expr::Field(base, member) = &recv.node
-            && self.local_value_enum_variant_target(base, member).is_some()
-        {
-            self.lower_expr_to_operand(recv, scope, out)
-        } else {
-            let recv_place = self.lower_expr_to_place(recv, scope, out);
-            bir::Operand::place(recv_place, bir::OwnershipFact::Borrow, false)
-        };
-
-        let (mut arg_operands, binding) =
-            match self.bind_declared_args(&format!("method `{name}`"), declared, args, scope, out) {
-                Ok(bound) => bound,
-                Err(description) => {
-                    return self.unsupported_operand(description, scope, hir_span_value, out);
-                }
-            };
-
-        // The receiver is `args[0]` and is never spliced, so it is always a single-value element.
-        let mut call_args = Vec::with_capacity(arg_operands.len() + 1);
-        call_args.push(bir::ArgumentElement::One(receiver_operand));
-        call_args.append(&mut arg_operands);
-        let ty = self.resolve_ty(span);
-        self.push_call_temp(
-            bir::Callee::Method(bir::MethodTarget {
-                name: name.to_string(),
-                type_args: resolved_type_args,
-                binding,
-            }),
-            call_args,
-            ty,
-            scope,
-            hir_span_value,
-            false,
-            out,
-        )
-    }
-
-    /// Lower a list literal, including spread entries, into a [`bir::AggregateKind::List`] aggregate.
-    ///
-    /// Elements are lowered in written source order, so a spread source's evaluation is interleaved with the fixed
-    /// elements around it exactly as written. A spread contributes one [`bir::ArgumentElement::Spread`] whose
-    /// length is a runtime fact; surrounding fixed elements keep their positions relative to it.
-    fn lower_list_literal(
-        &mut self,
-        entries: &[ast::ListEntry],
-        span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let hir_span_value = hir_span(span);
-        let mut elements = Vec::with_capacity(entries.len());
-        for entry in entries {
-            match entry {
-                ast::ListEntry::Element(item) => {
-                    elements.push(bir::ArgumentElement::One(self.lower_expr_to_operand(item, scope, out)));
-                }
-                ast::ListEntry::Spread(source) => {
-                    elements.push(self.lower_spread_element(source, bir::SpreadKind::Sequence, scope, out));
-                }
-            }
-        }
-        let ty = self.resolve_ty(span);
-        self.record_runtime_requirement(AbiV0RuntimeRequirement::Allocator);
-        self.push_assign_temp(
-            bir::Rvalue::Aggregate(bir::AggregateKind::List, elements),
-            ty,
-            scope,
-            hir_span_value,
-            out,
-        )
-    }
-
-    /// Lower one spread source into a [`bir::ArgumentElement::Spread`].
-    ///
-    /// The source is read through the ordinary ownership path, so a spliced source carries the same
-    /// [`bir::OwnershipFact`]/last-use discipline as any other read. That fact is recorded on the spread itself
-    /// rather than inferred from the surrounding aggregate or call, because a spliced source is consumed
-    /// differently from a single element: its contents are distributed into the surrounding list.
-    fn lower_spread_element(
-        &mut self,
-        source: &ast::Spanned<ast::Expr>,
-        kind: bir::SpreadKind,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::ArgumentElement {
-        let operand = self.lower_expr_to_operand(source, scope, out);
-        bir::ArgumentElement::Spread(bir::SpreadElement { source: operand, kind })
-    }
-
-    /// Lower a tuple or set literal to a [`bir::Rvalue::Aggregate`], recording an
-    /// [`AbiV0RuntimeRequirement::Allocator`] requirement for lists and sets specifically (list/set construction
-    /// always allocates; tuples do not).
-    fn lower_aggregate(
-        &mut self,
-        kind: bir::AggregateKind,
-        items: &[ast::Spanned<ast::Expr>],
-        span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let hir_span_value = hir_span(span);
-        let operands: Vec<bir::Operand> = items
-            .iter()
-            .map(|item| self.lower_expr_to_operand(item, scope, out))
-            .collect();
-        let ty = self.resolve_ty(span);
-        if matches!(kind, bir::AggregateKind::List | bir::AggregateKind::Set) {
-            self.record_runtime_requirement(AbiV0RuntimeRequirement::Allocator);
-        }
-        self.push_assign_temp(
-            bir::Rvalue::Aggregate(kind, fixed_elements(operands)),
-            ty,
-            scope,
-            hir_span_value,
-            out,
-        )
-    }
-
-    /// Lower a dict literal `{k: v, ...}` to a [`bir::Rvalue::Dict`], one entry per source entry, in written order.
-    ///
-    /// Keys and values are lowered in written order, key before value, because both are arbitrary expressions
-    /// whose evaluation order is source-observable. A `**source` spread contributes one
-    /// [`bir::DictEntry::Spread`] in written position; entries take effect in order and a later entry overwrites an
-    /// earlier one with the same key, which is what makes `{**base, "x": 1}` well defined.
-    fn lower_dict(
-        &mut self,
-        entries: &[ast::DictEntry],
-        span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let hir_span_value = hir_span(span);
-        let mut lowered = Vec::with_capacity(entries.len());
-        for entry in entries {
-            match entry {
-                ast::DictEntry::Pair(key, value) => {
-                    let key_operand = self.lower_expr_to_operand(key, scope, out);
-                    let value_operand = self.lower_expr_to_operand(value, scope, out);
-                    lowered.push(bir::DictEntry::Pair(key_operand, value_operand));
-                }
-                ast::DictEntry::Spread(source) => {
-                    // Reuse the shared spread lowering so the two construction sites cannot drift.
-                    let bir::ArgumentElement::Spread(spread) =
-                        self.lower_spread_element(source, bir::SpreadKind::Mapping, scope, out)
-                    else {
-                        return self.unsupported_operand("dict spread entry".to_string(), scope, hir_span_value, out);
-                    };
-                    lowered.push(bir::DictEntry::Spread(spread));
-                }
-            }
-        }
-        let ty = self.resolve_ty(span);
-        self.record_runtime_requirement(AbiV0RuntimeRequirement::Allocator);
-        self.push_assign_temp(bir::Rvalue::Dict(lowered), ty, scope, hir_span_value, out)
-    }
-
-    /// Lower an f-string `f"...{expr}...{expr!r}..."` to a [`bir::Rvalue::Format`]. Literal text chunks are
-    /// carried through verbatim; each embedded expression is lowered through the same
-    /// [`Self::lower_expr_to_operand`] path as any other read, so ownership facts and last-use tracking apply to
-    /// f-string interpolations exactly like any other expression use. Mirrors the existing Rust-emission backend's
-    /// dedicated `Format` node (`src/backend/ir/lower/expr/mod.rs`) rather than desugaring into a helper call --
-    /// see [`bir::Rvalue::Format`]'s own docs for why this needed its own `Rvalue` shape.
-    ///
-    /// Building the formatted string always allocates and always needs the `fstring` runtime helper
-    /// (`incan_stdlib::strings::fstring`, the function the existing Rust-emission backend's `Format` node itself
-    /// compiles down to -- see `src/backend/ir/emit/expressions/format.rs`), so both requirements are recorded
-    /// unconditionally here, the same way [`Self::lower_binary_from_operands`] records requirements for its own
-    /// compiler-owned string helpers.
-    fn lower_fstring(
-        &mut self,
-        parts: &[ast::FStringPart],
-        span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let hir_span_value = hir_span(span);
-        let ir_parts: Vec<bir::FormatPart> = parts
-            .iter()
-            .map(|part| match part {
-                ast::FStringPart::Literal(s) => bir::FormatPart::Literal(s.clone()),
-                ast::FStringPart::Expr { expr, format } => {
-                    let operand = self.lower_expr_to_operand(expr, scope, out);
-                    let style = match format {
-                        ast::FStringFormat::Display => bir::FormatStyle::Display,
-                        ast::FStringFormat::Debug => bir::FormatStyle::Debug,
-                    };
-                    bir::FormatPart::Expr { operand, style }
-                }
-            })
-            .collect();
-        self.record_runtime_requirement(AbiV0RuntimeRequirement::RuntimeHelper("fstring".to_string()));
-        self.record_runtime_requirement(AbiV0RuntimeRequirement::Allocator);
-        let ty = self.resolve_ty(span);
-        self.push_assign_temp(bir::Rvalue::Format(ir_parts), ty, scope, hir_span_value, out)
-    }
-
-    /// Lower `base[start:end:step]` (each component independently optional) into a value read through a
-    /// [`bir::PlaceElem::Slice`] projection, mirroring how `Expr::Index` builds an `[index]`-projected place read
-    /// in [`Self::lower_expr_to_operand`] (including that same arm's index-before-base evaluation order, extended
-    /// here to start-then-end-then-step-then-base).
-    fn lower_slice(
-        &mut self,
-        base: &ast::Spanned<ast::Expr>,
-        slice: &ast::SliceExpr,
-        span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let start = slice
-            .start
-            .as_ref()
-            .map(|e| Box::new(self.lower_expr_to_operand(e, scope, out)));
-        let end = slice
-            .end
-            .as_ref()
-            .map(|e| Box::new(self.lower_expr_to_operand(e, scope, out)));
-        let step = slice
-            .step
-            .as_ref()
-            .map(|e| Box::new(self.lower_expr_to_operand(e, scope, out)));
-        let mut place = self.lower_expr_to_place(base, scope, out);
-        place.projection.push(bir::PlaceElem::Slice { start, end, step });
-        let ty = self.resolve_ty(span);
-        let (fact, last_use) = self.ownership_fact_for_place(&place, &ty);
-        bir::Operand::place(place, fact, last_use)
-    }
-
-    /// Lower `expr?` (`ast::Expr::Try`) into a single [`bir::StatementKind::TryPropagate`] primitive rather than
-    /// decomposing it into explicit `is_err`/`unwrap`-shaped calls -- see that variant's own docs for the full
-    /// rationale (it mirrors the same #653-criterion-3 compiler-owned-primitive treatment as
-    /// [`bir::Callee::Helper`], standing in for what the existing Rust-emission backend defers entirely to Rust's
-    /// native `?` operator).
-    fn lower_try(
-        &mut self,
-        inner: &ast::Spanned<ast::Expr>,
-        outer_span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let hir_span_value = hir_span(outer_span);
-        let operand_result_type = self.resolve_ty(inner.span);
-        let error_routing = match (
-            result_error_type(&operand_result_type),
-            result_error_type(&self.owner_return_type),
-        ) {
-            (Some(source_error_type), Some(destination_error_type)) if source_error_type == destination_error_type => {
-                bir::TryErrorRouting::SameType {
-                    error_type: source_error_type.clone(),
-                }
-            }
-            (Some(source_error_type), Some(destination_error_type)) => bir::TryErrorRouting::ConversionRequired {
-                source_error_type: source_error_type.clone(),
-                destination_error_type: destination_error_type.clone(),
-            },
-            _ => bir::TryErrorRouting::Unresolved,
-        };
-        let operand = self.lower_expr_to_operand(inner, scope, out);
-        let ty = self.resolve_ty(outer_span);
-        let destination = self.new_temp(ty.clone(), scope, hir_span_value);
-        out.push(bir::Statement {
-            kind: bir::StatementKind::TryPropagate {
-                destination: bir::Place::from_local(destination),
-                operand,
-                error_routing,
-            },
-            span: hir_span_value,
-        });
-        self.temp_operand(destination, &ty)
-    }
-
-    /// Lower an `ast::Expr::Constructor` node by delegating to [`Self::lower_nominal_construction`].
-    ///
-    /// No stage of the current pipeline produces this AST variant: `P(x=1, y=2)` parses as an
-    /// `ast::Expr::Call` whose callee is a bare identifier, and `lower_call` recognises the construction from the
-    /// typechecker's recorded field binding. The arm is kept because the variant is still part of the AST contract,
-    /// and it delegates rather than duplicating the lowering so a future producer cannot reach a second, divergent
-    /// construction path.
-    fn lower_constructor(
-        &mut self,
-        name: &str,
-        args: &[ast::CallArg],
-        span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        self.lower_nominal_construction(name, args, span, scope, out)
-    }
-
-    // ---- Async surface (#1164) ----
-
-    /// Lower an `ast::Expr::Surface` node, accepting only the async pair this issue owns.
-    ///
-    /// Dispatch is on the surface **key**, not the payload shape. `SurfaceExprPayload::PrefixUnary` is generic over
-    /// any prefix soft keyword and `await` merely happens to be the only one registered today, so matching the
-    /// payload alone would silently accept a future prefix keyword as an await. The typechecker
-    /// (`check_expr/mod.rs`) and the existing Rust-emission backend (`backend/ir/lower/expr/mod.rs`) both dispatch
-    /// on the key/payload pair for exactly this reason.
-    ///
-    /// Every other payload -- the scoped-DSL surface nodes -- keeps its existing named refusal. Those reach this
-    /// module only when a caller skips the desugar pass the legacy pipeline runs first, and they belong to the Body
-    /// IR input-contract issue, not to this one.
-    fn lower_surface_expr(
-        &mut self,
-        surface: &ast::SurfaceExpr,
-        span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        match (&surface.key, &surface.payload) {
-            (SurfaceFeatureKey::SoftKeyword(KeywordId::Await), ast::SurfaceExprPayload::PrefixUnary(awaited)) => {
-                self.lower_await(awaited, span, scope, out)
-            }
-            (
-                SurfaceFeatureKey::ScopedDslSurface {
-                    dependency_key,
-                    descriptor_key,
-                },
-                ast::SurfaceExprPayload::RaceFor(race),
-            ) if dependency_key == "std.async" && descriptor_key == "race_for" => {
-                self.lower_race_for(race, span, scope, out)
-            }
-            (_, payload) => self.unsupported_operand(surface_expr_label(payload), scope, hir_span(span), out),
-        }
-    }
-
-    /// Lower `await expr` into a [`bir::StatementKind::Await`] suspension point.
-    ///
-    /// The awaited operand is read through the ordinary ownership path, so the suspension carries the same
-    /// [`bir::OwnershipFact`]/last-use discipline as any other read. The resumed value lands in a fresh temporary,
-    /// which is what makes the suspension's destination explicit rather than implied by the surrounding statement.
-    ///
-    /// Records [`AbiV0RuntimeRequirement::AsyncRuntime`] on the enclosing body so a consumer reads the requirement
-    /// off the body it applies to instead of re-deriving it from the program's imports and declaration modifiers.
-    fn lower_await(
-        &mut self,
-        awaited: &ast::Spanned<ast::Expr>,
-        span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let hir_span_value = hir_span(span);
-        let operand = self.lower_expr_to_operand(awaited, scope, out);
-        self.record_runtime_requirement(AbiV0RuntimeRequirement::AsyncRuntime);
-        let ty = self.resolve_ty(span);
-        let destination = self.new_temp(ty.clone(), scope, hir_span_value);
-        out.push(bir::Statement {
-            kind: bir::StatementKind::Await {
-                destination: Some(bir::Place::from_local(destination)),
-                awaited: operand,
-            },
-            span: hir_span_value,
-        });
-        self.temp_operand(destination, &ty)
-    }
-
-    /// Lower `race for value:` into a [`bir::StatementKind::Race`].
-    ///
-    /// Each arm's awaitable is lowered into the enclosing block *before* any arm body, which is what makes "every
-    /// awaitable is evaluated before selection" observable in the statement sequence rather than a claim in prose.
-    /// Each arm then gets its own scope and its own binding local: the source spells one shared name, but arms
-    /// re-scope it and can resolve it to different types, so one local per arm is the faithful shape.
-    ///
-    /// An arm body containing an unsupported construct keeps its own `Unsupported` node *inside* the represented
-    /// race rather than collapsing the whole expression, so a consumer loses only the construct it cannot handle.
-    fn lower_race_for(
-        &mut self,
-        race: &ast::RaceForExpr,
-        span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let hir_span_value = hir_span(span);
-
-        // Selection observes every awaitable, so all of them are evaluated first, in source order, into the
-        // enclosing block. Only the winning arm's body runs, so arm bodies are lowered into their own blocks below.
-        let mut awaitables = Vec::with_capacity(race.arms.len());
-        for arm in &race.arms {
-            awaitables.push(self.lower_expr_to_operand(&arm.awaitable, scope, out));
-        }
-        self.record_runtime_requirement(AbiV0RuntimeRequirement::AsyncRuntime);
-
-        let mut arms = Vec::with_capacity(race.arms.len());
-        for (arm, awaitable) in race.arms.iter().zip(awaitables) {
-            let arm_scope = self.new_scope(Some(scope), hir_span_value);
-            // The arm binds the *awaited output* type, which only the typechecker computes: `Awaitable[T]` binds
-            // `T`, `JoinHandle[T]` binds `Result[T, TaskJoinError]`. The awaitable's own type would be wrong.
-            let binding_ty = self
-                .type_info
-                .race_arm_binding_type(arm.awaitable.span)
-                .map(semantic_type_from_resolved)
-                .unwrap_or(IncanType::Unknown);
-            // Snapshot the whole binding environment before the arm, not just the shared race binding. A block arm
-            // lowers ordinary statements, and every `x = ...` in it declares a local that
-            // `declare_new_local_with_reads` installs into `self.bindings`. Restoring only `race.binding`
-            // would leave those arm-locals visible to later arms and to code after the race, so a trailing
-            // read of a name an arm happened to shadow would silently resolve to the arm's local.
-            // `insert_scope_drops` handles the *drop* obligation; it does not touch name resolution, which
-            // is what this restores.
-            let enclosing_bindings = self.bindings.clone();
-            let reads = match &arm.body {
-                ast::RaceForBody::Expr(expr) => count_reads_in_expr(&race.binding, &expr.node),
-                ast::RaceForBody::Block(stmts) => count_reads_in_stmts(&race.binding, stmts),
-            };
-            let binding =
-                self.declare_new_local_with_reads(race.binding.clone(), binding_ty, arm_scope, hir_span_value, reads);
-
-            let mut arm_stmts = Vec::new();
-            let result = match &arm.body {
-                ast::RaceForBody::Expr(expr) => self.lower_expr_to_operand(expr, arm_scope, &mut arm_stmts),
-                ast::RaceForBody::Block(stmts) => {
-                    self.lower_race_arm_block(stmts, arm.awaitable.span, arm_scope, &mut arm_stmts)
-                }
-            };
-            self.insert_scope_drops(&mut arm_stmts, arm_scope);
-
-            // Every name an arm bound -- its winner binding and any local its block body declared -- is scoped to
-            // that arm, exactly like a closure body's. Code after the race, and each later arm, must keep resolving
-            // every name to whatever it meant outside.
-            self.bindings = enclosing_bindings;
-
-            arms.push(bir::RaceArm {
-                awaitable,
-                binding,
-                body: bir::Block {
-                    scope: arm_scope,
-                    stmts: arm_stmts,
-                },
-                result,
-            });
-        }
-
-        let ty = self.resolve_ty(span);
-        let destination = self.new_temp(ty.clone(), scope, hir_span_value);
-        out.push(bir::Statement {
-            kind: bir::StatementKind::Race {
-                destination: Some(bir::Place::from_local(destination)),
-                arms,
-            },
-            span: hir_span_value,
-        });
-        self.temp_operand(destination, &ty)
-    }
-
-    /// Lower a race arm's block body, whose value is its trailing expression statement.
-    ///
-    /// That trailing-expression convention is the source contract the typechecker already applies, so lowering
-    /// matches `check_race_arm_block_body` exactly, including its two non-expression cases: an empty block and a
-    /// block whose last statement is not an expression both produce `Unit`, the same type the checker assigns them.
-    /// Refusing either would make a program the source language accepts unrepresentable, and the established
-    /// precedent for a valueless block arm is [`Self::lower_match`]'s own block body.
-    fn lower_race_arm_block(
-        &mut self,
-        stmts: &[ast::Spanned<ast::Statement>],
-        arm_span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let Some((last, leading)) = stmts.split_last() else {
-            return bir::Operand::Constant(bir::Constant::Unit);
-        };
-        for (index, stmt) in leading.iter().enumerate() {
-            self.lower_stmt_into(stmt, &stmts[index + 1..], scope, out);
-        }
-        match &last.node {
-            ast::Statement::Expr(expr) => self.lower_expr_to_operand(expr, scope, out),
-            _ => {
-                let _ = arm_span;
-                self.lower_stmt_into(last, &[], scope, out);
-                bir::Operand::Constant(bir::Constant::Unit)
-            }
-        }
-    }
-
-    // ---- Closures and partial callables (#1101 bucket B4) ----
-
-    /// Lower a closure literal `(params) => expr` into a [`bir::Rvalue::Closure`].
-    ///
-    /// Body IR must represent captures explicitly rather than deferring to a consuming backend's own closure syntax
-    /// to auto-capture (see this module's docs and #1101's B4 pre-intake), so this: (1) statically determines every
-    /// free variable the closure body reads via [`free_vars_in_closure_body`]; (2) reads each one exactly once, at
-    /// this closure-creation site, through the same [`Self::ownership_fact_for_place`] path any other read in this
-    /// body uses, recording the result as this closure's `captured_operands`; (3) declares a fresh
-    /// [`bir::LocalOrigin::Captured`] local per capture plus one [`bir::LocalOrigin::Parameter`] local per declared
-    /// parameter, shadowing (and restoring afterward) any outer binding of the same name, so the closure body's own
-    /// reads resolve to its own bound copy rather than silently reading through to the enclosing scope; then (4)
-    /// lowers the body expression under those bindings. The restore step is what makes this different from every
-    /// other nested block this file lowers -- ordinary nested blocks (`if`/`loop` bodies) let a shadowing binding
-    /// leak forward in `self.bindings` with no restore, which is harmless for straight-line control flow but would
-    /// be wrong here: code lexically after the closure literal must keep resolving the shadowed name to the
-    /// *enclosing* variable, not to the closure's own captured copy.
-    fn lower_closure(
-        &mut self,
-        params: &[ast::Spanned<ast::Param>],
-        body_expr: &ast::Spanned<ast::Expr>,
-        expr_span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let hir_span_value = hir_span(expr_span);
-        let closure_scope = self.new_scope(Some(scope), hir_span_value);
-
-        // ---- Capture every free variable exactly once, at this closure-creation site ----
-        let free_names = free_vars_in_closure_body(params, body_expr);
-        let mut captured_operands = Vec::with_capacity(free_names.len());
-        let mut capture_locals = Vec::with_capacity(free_names.len());
-        let mut saved_bindings: Vec<(String, Option<bir::LocalId>)> = Vec::new();
-        for name in &free_names {
-            // A free name lowering cannot resolve to a tracked outer local (e.g. a module-level `const`) is not
-            // captured -- the closure body's own `Self::local_for_name` lookup synthesizes an `External` reference
-            // for it exactly like anywhere else, since there is nothing meaningful to read-and-rebind.
-            let Some(&outer_local) = self.bindings.get(name) else {
-                continue;
-            };
-            let outer_ty = self.locals[outer_local.index()].ty.clone();
-            let outer_place = bir::Place::from_local(outer_local);
-            let (fact, last_use) = self.ownership_fact_for_place(&outer_place, &outer_ty);
-            captured_operands.push(bir::Operand::place(outer_place, fact, last_use));
-
-            let total_reads = count_reads_in_expr(name, &body_expr.node);
-            let capture_local =
-                self.declare_new_local_with_reads(name.clone(), outer_ty, closure_scope, hir_span_value, total_reads);
-            self.locals[capture_local.index()].origin = bir::LocalOrigin::Captured;
-            capture_locals.push(capture_local);
-            saved_bindings.push((name.clone(), Some(outer_local)));
-        }
-
-        // ---- Bind the closure's own parameters, shadowing any outer binding of the same name ----
-        let param_types = self.closure_param_types(params, expr_span);
-        let mut closure_param_locals = Vec::with_capacity(params.len());
-        for (param, ty) in params.iter().zip(param_types) {
-            let previous = self.bindings.get(&param.node.name).copied();
-            let total_reads = count_reads_in_expr(&param.node.name, &body_expr.node);
-            let local = self.declare_new_local_with_reads(
-                param.node.name.clone(),
-                ty.clone(),
-                closure_scope,
-                hir_span(param.span),
-                total_reads,
-            );
-            self.locals[local.index()].origin = bir::LocalOrigin::Parameter;
-            closure_param_locals.push(local);
-            saved_bindings.push((param.node.name.clone(), previous));
-        }
-
-        let mut closure_params = Vec::with_capacity(params.len());
-        for (param, local) in params.iter().zip(closure_param_locals) {
-            let ty = self.locals[local.index()].ty.clone();
-            closure_params.push(bir::CallableParam {
-                local,
-                name: param.node.name.clone(),
-                ty,
-                span: hir_span(param.span),
-                default: self.lower_callable_default(param.node.default.as_ref(), closure_scope),
-            });
-        }
-
-        // ---- Lower the body under the closure's own bindings, then restore the enclosing scope's ----
-        let mut body_stmts = Vec::new();
-        let result = self.lower_expr_to_operand(body_expr, closure_scope, &mut body_stmts);
-        for (name, previous) in saved_bindings {
-            match previous {
-                Some(local) => {
-                    self.bindings.insert(name, local);
-                }
-                None => {
-                    self.bindings.remove(&name);
-                }
-            }
-        }
-
-        let closure_body = bir::ClosureBody {
-            capture_locals,
-            stmts: body_stmts,
-            result,
-        };
-        let ty = self.resolve_ty(expr_span);
-        self.push_assign_temp(
-            bir::Rvalue::Closure {
-                params: closure_params,
-                captured_operands,
-                body: Box::new(closure_body),
-            },
-            ty,
-            scope,
-            hir_span_value,
-            out,
-        )
-    }
-
-    /// Resolve each of a closure literal's parameter types from the typechecker's resolved callable type at the
-    /// closure's own span, falling back to [`IncanType::Unknown`] per parameter when unavailable or of mismatched
-    /// length. Mirrors the existing Rust-emission backend's own `recorded_param_types` fallback
-    /// (`src/backend/ir/lower/expr/mod.rs`), minus that backend's additional Rust-display-exact override, which is
-    /// meaningful only for concrete Rust closure syntax, not this target-agnostic model.
-    fn closure_param_types(&self, params: &[ast::Spanned<ast::Param>], expr_span: ast::Span) -> Vec<IncanType> {
-        let resolved = self.type_info.expr_type(expr_span).and_then(|ty| match ty {
-            ResolvedType::Function(callable_params, _) => Some(
-                callable_params
-                    .iter()
-                    .map(|p| semantic_type_from_resolved(&p.ty))
-                    .collect::<Vec<_>>(),
-            ),
-            _ => None,
-        });
-        match resolved {
-            Some(types) if types.len() == params.len() => types,
-            _ => vec![IncanType::Unknown; params.len()],
-        }
-    }
-
-    /// Lower a partial callable preset expression (`partial Target(name=value, ...)`) into the same
-    /// [`bir::Rvalue::Closure`] shape a closure literal produces, mirroring how the existing Rust-emission backend
-    /// already desugars a partial application into a synthesized closure that forwards the still-missing arguments
-    /// into a call (`src/backend/ir/lower/expr/mod.rs`'s `ast::Expr::Partial` arm) -- see #1101's B4 pre-intake.
-    /// Partial construction currently supports only a bare top-level function-name `target` whose full parameter list
-    /// the typechecker resolved. General Body IR calls still distinguish named functions from local callable values
-    /// and record local supplied-parameter slots (see [`Self::lower_call`]). A method-shaped partial target from
-    /// `partial recv.method(...)`, explicit type arguments, or a target with an unnamed parameter lowers to an
-    /// explicit unsupported placeholder instead.
-    ///
-    /// Preset values (`partial.args`) are lowered once each, at the partial-creation site -- exactly like an
-    /// ordinary call argument, not deduplicated per free-variable name the way [`Self::lower_closure`]'s captures
-    /// are -- and folded into the synthesized closure's own `captured_operands`. Every declared target parameter
-    /// remains a closure parameter in declaration order. A preset parameter records
-    /// [`bir::CallableParamDefault::PartialPreset`], while an unpresetted target default retains its distinct
-    /// source-default contract: a deferred [`bir::CallableParamDefault::Source`] computation only when it has
-    /// usable type facts, otherwise an original-span refusal. Positional local calls skip only preset parameters;
-    /// [`Self::lower_call`] records the supplied declaration slots rather than pretending the complete callable
-    /// surface is a residual function type.
-    ///
-    /// `Expr::Partial` uses this same full callable surface through `local_partial_params`; module-level partial
-    /// declarations intentionally keep their existing full-signature-plus-preset-metadata projection for backend
-    /// and export consumers. A compound-assignment-style mutation of a captured preset from inside a nested closure
-    /// is out of scope here in the same way [`Self::lower_closure`]'s own docs note for ordinary closures.
-    fn lower_partial(
-        &mut self,
-        partial: &ast::PartialExpr,
-        expr_span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let hir_span_value = hir_span(expr_span);
-        let ast::Expr::Ident(target_name) = &partial.target.node else {
-            return self.unsupported_operand(
-                "partial callable with a non-function-name target".to_string(),
-                scope,
-                hir_span_value,
-                out,
-            );
-        };
-        if !partial.type_args.is_empty() {
-            return self.unsupported_operand(
-                "partial callable with explicit type arguments".to_string(),
-                scope,
-                hir_span_value,
-                out,
-            );
-        }
-        let Some(binding) = self.type_info.declarations.function_bindings.get(target_name).cloned() else {
-            return self.unsupported_operand(
-                "partial callable target with no resolvable top-level function signature".to_string(),
-                scope,
-                hir_span_value,
-                out,
-            );
-        };
-        if binding
-            .params
-            .iter()
-            .any(|param| param.name.is_none() || param.kind != ast::ParamKind::Normal)
-        {
-            return self.unsupported_operand(
-                "partial callable target with an unnamed or rest parameter".to_string(),
-                scope,
-                hir_span_value,
-                out,
-            );
-        }
-        let target_name = target_name.clone();
-        let direct_call_id = self
-            .local_function_declarations
-            .get(&target_name)
-            .and_then(|candidates| match candidates.as_slice() {
-                [target_span] => Some(CompilerNodeId::declaration_span(
-                    self.module_identity,
-                    target_span.start,
-                    target_span.end,
-                )),
-                _ => None,
-            });
-        let target_default_sources = self.function_default_sources.get(&target_name).cloned();
-        let closure_scope = self.new_scope(Some(scope), hir_span_value);
-
-        // ---- Lower each preset value once, at the partial-creation site, as a captured operand ----
-        let mut captured_operands = Vec::with_capacity(partial.args.len());
-        let mut capture_locals = Vec::with_capacity(partial.args.len());
-        let mut preset_lookup: HashMap<String, bir::LocalId> = HashMap::with_capacity(partial.args.len());
-        let mut saved_bindings = Vec::with_capacity(binding.params.len() + partial.args.len());
-        for arg in &partial.args {
-            let value_ty = self.resolve_ty(arg.value.span);
-            let operand = self.lower_expr_to_operand(&arg.value, scope, out);
-            captured_operands.push(operand);
-            let capture_name = format!("__partial_preset_{}", arg.name);
-            let previous = self.bindings.get(&capture_name).copied();
-            let capture_local =
-                self.declare_new_local_with_reads(capture_name.clone(), value_ty, closure_scope, hir_span_value, 1);
-            self.locals[capture_local.index()].origin = bir::LocalOrigin::Captured;
-            capture_locals.push(capture_local);
-            preset_lookup.insert(arg.name.clone(), capture_local);
-            saved_bindings.push((capture_name, previous));
-        }
-
-        // ---- Every target parameter stays on the closure surface; presets become overrideable defaults ----
-        let mut closure_params = Vec::new();
-        let mut call_arg_locals = Vec::with_capacity(binding.params.len());
-        for (index, param) in binding.params.iter().enumerate() {
-            let Some(param_name) = &param.name else {
-                return self.unsupported_operand(
-                    "partial callable target with an unnamed parameter".to_string(),
-                    scope,
-                    hir_span_value,
-                    out,
-                );
-            };
-            let ty = semantic_type_from_resolved(&param.ty);
-            let previous = self.bindings.get(param_name).copied();
-            let local =
-                self.declare_new_local_with_reads(param_name.clone(), ty.clone(), closure_scope, hir_span_value, 1);
-            self.locals[local.index()].origin = bir::LocalOrigin::Parameter;
-            let source_param = target_default_sources.as_ref().and_then(|params| params.get(index));
-            let default = match preset_lookup.get(param_name).copied() {
-                Some(capture) => bir::CallableParamDefault::PartialPreset { capture },
-                None => match source_param {
-                    Some(source_param) => self.lower_callable_default(source_param.default.as_ref(), closure_scope),
-                    None if param.has_default => bir::CallableParamDefault::Unsupported {
-                        span: hir_span_value,
-                        description: format!(
-                            "partial target {target_name} declares a default Body IR could not source"
-                        ),
-                    },
-                    None => bir::CallableParamDefault::Required,
-                },
-            };
-            closure_params.push(bir::CallableParam {
-                local,
-                name: param_name.clone(),
-                ty,
-                span: source_param.map_or(hir_span_value, |param| hir_span(param.param_span)),
-                default,
-            });
-            call_arg_locals.push(local);
-            saved_bindings.push((param_name.clone(), previous));
-        }
-
-        // ---- Synthesize the forwarding call as the closure's single-statement body ----
-        let mut body_stmts = Vec::new();
-        let call_args: Vec<bir::Operand> = call_arg_locals
-            .iter()
-            .zip(&binding.params)
-            .map(|(&local, param)| {
-                let ty = semantic_type_from_resolved(&param.ty);
-                let place = bir::Place::from_local(local);
-                let (fact, last_use) = self.ownership_fact_for_place(&place, &ty);
-                bir::Operand::place(place, fact, last_use)
-            })
-            .collect();
-        let ret_ty = semantic_type_from_resolved(&binding.return_type);
-        // The synthesized forwarding call supplies every declared parameter of the target, in declaration order:
-        // preset slots are filled from the captured locals and residual slots from the closure's own parameters.
-        let forwarding_binding = bir::ArgumentBinding::resolved_positional(call_args.len());
-        let result = self.push_call_temp(
-            bir::Callee::Function(bir::CallableTarget::Named(bir::NamedCallableTarget {
-                name: target_name,
-                direct_call_id,
-                builtin: None,
-                // A compiler-synthesized forwarding call has no source call site, so it has no spelling or import
-                // provenance to record; the declaration it forwards to is already named by `direct_call_id`.
-                canonical: None,
-                type_args: Vec::new(),
-                binding: forwarding_binding,
-            })),
-            fixed_elements(call_args),
-            ret_ty,
-            closure_scope,
-            hir_span_value,
-            false,
-            &mut body_stmts,
-        );
-
-        let closure_body = bir::ClosureBody {
-            capture_locals,
-            stmts: body_stmts,
-            result,
-        };
-
-        // ---- The synthesized closure's bindings are lexically private to it, not new outer bindings ----
-        for (name, previous) in saved_bindings.into_iter().rev() {
-            match previous {
-                Some(local) => {
-                    self.bindings.insert(name, local);
-                }
-                None => {
-                    self.bindings.remove(&name);
-                }
-            }
-        }
-
-        let ty = self.resolve_ty(expr_span);
-        self.push_assign_temp(
-            bir::Rvalue::Closure {
-                params: closure_params,
-                captured_operands,
-                body: Box::new(closure_body),
-            },
-            ty,
-            scope,
-            hir_span_value,
-            out,
-        )
-    }
-
-    /// Lower a `match` expression (`ast::Expr::Match`) into a single [`bir::Rvalue::Match`], mirroring the existing
-    /// Rust-emission backend's own `IrExprKind::Match { scrutinee, arms }` node -- see [`bir::Rvalue::Match`]'s docs
-    /// for why matching stays one structured node rather than being decomposed into a chain of `If` statements, and
-    /// [`bir::Pattern`]'s docs for the closed pattern vocabulary this mirrors and its two deliberate v0 gaps (no
-    /// union-type pattern narrowing, no RFC 021 field-alias resolution).
-    ///
-    /// Bails the whole expression to an explicit unsupported placeholder *before* lowering the scrutinee when any
-    /// arm's pattern contains a byte-string literal (the one pattern shape [`bir::Constant`] cannot represent --
-    /// see [`match_pattern_is_supported`]), mirroring [`Self::lower_binary`]'s "check before partially lowering"
-    /// precedent so an unrepresentable pattern never produces a partially-lowered `Rvalue::Match`.
-    fn lower_match(
-        &mut self,
-        subject: &ast::Spanned<ast::Expr>,
-        arms: &[ast::Spanned<ast::MatchArm>],
-        expr_span: ast::Span,
-        scope: bir::ScopeId,
-        out: &mut Vec<bir::Statement>,
-    ) -> bir::Operand {
-        let hir_span_value = hir_span(expr_span);
-        if arms
-            .iter()
-            .any(|arm| !match_pattern_is_supported(&arm.node.pattern.node))
-        {
-            return self.unsupported_operand(
-                "match arm with a byte-string literal pattern".to_string(),
-                scope,
-                hir_span_value,
-                out,
-            );
-        }
-
-        let scrutinee_ty = self.resolve_ty(subject.span);
-        let scrutinee_place = self.lower_expr_to_place(subject, scope, out);
-        // Always read as `Borrow` -- see `bir::Rvalue::Match::scrutinee`'s own docs for why the overall scrutinee
-        // read must not risk an unconditional move while individual pattern bindings below compute their own,
-        // more precise facts against projected places rooted at this same scrutinee.
-        let scrutinee_operand = bir::Operand::place(scrutinee_place.clone(), bir::OwnershipFact::Borrow, false);
-
-        let mut lowered_arms = Vec::with_capacity(arms.len());
-        for arm in arms {
-            let arm_span = hir_span(arm.span);
-            let arm_scope = self.new_scope(Some(scope), arm_span);
-
-            // ---- Lower the pattern, declaring one fresh arm-scoped local per distinct bound name ----
-            let mut seen: HashMap<String, bir::LocalId> = HashMap::new();
-            let mut saved_bindings: Vec<(String, Option<bir::LocalId>)> = Vec::new();
-            let pattern = self.lower_match_pattern(
-                &arm.node.pattern,
-                &scrutinee_ty,
-                &scrutinee_place,
-                arm_scope,
-                &arm.node,
-                &mut seen,
-                &mut saved_bindings,
-            );
-
-            // ---- Guard and body see this arm's own pattern bindings, shadowing any outer binding of the same name
-            // ----
-            let mut guard_stmts = Vec::new();
-            let guard = arm
-                .node
-                .guard
-                .as_ref()
-                .map(|g| self.lower_expr_to_operand(g, arm_scope, &mut guard_stmts));
-
-            let (body_stmts, result) = match &arm.node.body {
-                ast::MatchBody::Expr(e) => {
-                    let mut stmts = Vec::new();
-                    let result = self.lower_expr_to_operand(e, arm_scope, &mut stmts);
-                    (stmts, result)
-                }
-                ast::MatchBody::Block(block_stmts) => {
-                    let mut stmts = Vec::new();
-                    self.lower_block_into(block_stmts, arm_scope, &mut stmts);
-                    self.insert_scope_drops(&mut stmts, arm_scope);
-                    (stmts, bir::Operand::Constant(bir::Constant::Unit))
-                }
-            };
-
-            // ---- Restore the enclosing scope's bindings before moving on to the next (mutually exclusive) arm ----
-            for (name, previous) in saved_bindings {
-                match previous {
-                    Some(local) => {
-                        self.bindings.insert(name, local);
-                    }
-                    None => {
-                        self.bindings.remove(&name);
-                    }
-                }
-            }
-
-            lowered_arms.push(bir::MatchArm {
-                pattern,
-                guard_stmts,
-                guard,
-                body_stmts,
-                result,
-            });
-        }
-
-        let ty = self.resolve_ty(expr_span);
-        self.push_assign_temp(
-            bir::Rvalue::Match {
-                scrutinee: scrutinee_operand,
-                arms: lowered_arms,
-            },
-            ty,
-            scope,
-            hir_span_value,
-            out,
-        )
-    }
-
-    /// Recursively lower one source `ast::Pattern` node into a [`bir::Pattern`], declaring a fresh arm-scoped local
-    /// the first time a bound name is encountered and reusing it for any later `Or`-alternative occurrence of the
-    /// same name (`seen`) -- Incan's typechecker (RFC 071) requires every alternative of an `A(x) | B(x)` pattern to
-    /// bind an identical name/type set, so Rust's own single shared binding slot per name is the correct target
-    /// shape, not one local per occurrence. `saved_bindings` accumulates `(name, previous_local)` pairs so
-    /// [`Self::lower_match`] can restore `self.bindings` to the enclosing scope once this arm's guard/body have
-    /// both been lowered, the same save/restore shape [`Self::lower_closure`] already uses around its own
-    /// params/captures.
-    ///
-    /// `place` is the (possibly already-projected) scrutinee place this pattern node corresponds to; each
-    /// recursive call into a `Tuple`/`Struct`/`Enum` sub-pattern extends it with one more
-    /// [`bir::PlaceElem::Field`] projection -- named for a struct field, or the zero-based positional index as a
-    /// string for a tuple/enum-variant positional field, mirroring [`Self::lower_tuple_unpack`]'s own tuple-element
-    /// projection convention (`.0`/`.1` Rust tuple-field-access spelling) rather than inventing a second one.
-    ///
-    /// `expected_ty` is the best available type for this pattern node: propagated through [`Self::lower_match`]'s
-    /// own `Self::resolve_ty` call on the scrutinee for the root pattern, and through
-    /// [`tuple_element_types`] for `Tuple` sub-patterns (both already-established sources elsewhere in this file);
-    /// a `Struct`/`Enum` constructor pattern's own fields fall back to [`IncanType::Unknown`] per field, since
-    /// resolving a model/class/enum-variant's real field types would mean rebuilding the existing Rust-emission
-    /// backend's own field-type-projection machinery (`constructor_field_types_for_pattern` in
-    /// `src/backend/ir/lower/expr/patterns.rs`), which this bucket deliberately does not mirror -- see
-    /// [`bir::Pattern`]'s own docs.
-    #[allow(clippy::too_many_arguments)]
-    fn lower_match_pattern(
-        &mut self,
-        pattern: &ast::Spanned<ast::Pattern>,
-        expected_ty: &IncanType,
-        place: &bir::Place,
-        arm_scope: bir::ScopeId,
-        arm: &ast::MatchArm,
-        seen: &mut HashMap<String, bir::LocalId>,
-        saved_bindings: &mut Vec<(String, Option<bir::LocalId>)>,
-    ) -> bir::Pattern {
-        let span = hir_span(pattern.span);
-        match &pattern.node {
-            ast::Pattern::Wildcard => bir::Pattern::Wildcard,
-            ast::Pattern::Binding(name) => {
-                let local = match seen.get(name) {
-                    Some(&local) => local,
-                    None => {
-                        let total_reads = count_reads_in_match_arm(name, arm);
-                        let previous = self.bindings.get(name).copied();
-                        let local = self.declare_new_local_with_reads(
-                            name.clone(),
-                            expected_ty.clone(),
-                            arm_scope,
-                            span,
-                            total_reads,
-                        );
-                        seen.insert(name.clone(), local);
-                        saved_bindings.push((name.clone(), previous));
-                        local
-                    }
-                };
-                let (fact, last_use) = self.ownership_fact_for_place(place, expected_ty);
-                bir::Pattern::Var(bir::PatternBinding { local, fact, last_use })
-            }
-            // `match_pattern_is_supported` has already ruled out the one shape `lower_literal` cannot represent
-            // (a byte-string literal) for every arm in this match before `Self::lower_match` calls this method at
-            // all, so the `None` case here is unreachable in practice; `Constant::Unit` is a harmless, structurally
-            // valid fallback rather than a panic if that invariant is ever violated.
-            ast::Pattern::Literal(lit) => bir::Pattern::Literal(lower_literal(lit).unwrap_or(bir::Constant::Unit)),
-            ast::Pattern::Tuple(items) => {
-                let element_types = tuple_element_types(expected_ty, items.len());
-                let fields = items
-                    .iter()
-                    .zip(element_types.iter())
-                    .enumerate()
-                    .map(|(index, (item, element_ty))| {
-                        let mut field_place = place.clone();
-                        field_place.projection.push(bir::PlaceElem::Field(index.to_string()));
-                        self.lower_match_pattern(item, element_ty, &field_place, arm_scope, arm, seen, saved_bindings)
-                    })
-                    .collect();
-                bir::Pattern::Tuple(fields)
-            }
-            ast::Pattern::Constructor(name, args) => {
-                // Preserve exact source-local pattern targets instead of asking the executor to recover a
-                // declaration from the printed constructor spelling. The direct profile accepts only canonical
-                // named fields of a plain model; every other structurally lowered constructor remains the
-                // name-only fallback below and is visibly refused by replacement execution.
-                if let Some(declaration) = self.local_nominal_declarations.get(name)
-                    && matches!(expected_ty, IncanType::Named(type_name) if type_name == name)
-                    && args.iter().all(|arg| matches!(arg, ast::PatternArg::Named(_, _)))
-                {
-                    let fields = args
-                        .iter()
-                        .filter_map(|arg| match arg {
-                            ast::PatternArg::Named(field, pat) => {
-                                let mut field_place = place.clone();
-                                field_place.projection.push(bir::PlaceElem::Field(field.clone()));
-                                Some((
-                                    field.clone(),
-                                    self.lower_match_pattern(
-                                        pat,
-                                        &IncanType::Unknown,
-                                        &field_place,
-                                        arm_scope,
-                                        arm,
-                                        seen,
-                                        saved_bindings,
-                                    ),
-                                ))
-                            }
-                            ast::PatternArg::Positional(_) => None,
-                        })
-                        .collect();
-                    return bir::Pattern::Nominal {
-                        target: bir::NominalPatternTarget {
-                            direct_declaration_id: declaration.direct_declaration_id.clone(),
-                            name: declaration.name.clone(),
-                        },
-                        fields,
-                    };
-                }
-
-                if let Some((enum_name, variant_name)) = name.rsplit_once("::").or_else(|| name.rsplit_once('.'))
-                    && args.is_empty()
-                    && matches!(expected_ty, IncanType::Named(type_name) if type_name == enum_name)
-                    && let Some(declaration) = self.local_fieldless_enum_declarations.get(enum_name)
-                    && let Some(variant) = declaration.variants.iter().find(|variant| variant.name == variant_name)
-                {
-                    return bir::Pattern::FieldlessEnumVariant(bir::FieldlessEnumVariantTarget {
-                        enum_declaration_id: declaration.direct_declaration_id.clone(),
-                        variant_declaration_id: variant.direct_declaration_id.clone(),
-                        enum_name: declaration.name.clone(),
-                        variant_name: variant.name.clone(),
-                    });
-                }
-
-                if let Some(variant) = result_variant_kind(name)
-                    && let Some((ok_type, error_type)) = result_type_parts(expected_ty)
-                    && args.len() == 1
-                    && let [ast::PatternArg::Positional(payload)] = args.as_slice()
-                {
-                    let payload_type = match variant {
-                        bir::ResultVariantKind::Ok => ok_type,
-                        bir::ResultVariantKind::Err => error_type,
-                    };
-                    let lowered_payload =
-                        self.lower_match_pattern(payload, payload_type, place, arm_scope, arm, seen, saved_bindings);
-                    return bir::Pattern::Result {
-                        variant,
-                        fields: vec![lowered_payload],
-                    };
-                }
-
-                // Mirrors the existing Rust-emission backend's own `lower_pattern` (non-union-aware) mapping
-                // exactly: a mix of named and positional arguments (unusual, likely non-representative source)
-                // still lowers every sub-pattern's own bindings for side effects, but only the named fields survive
-                // into the constructed `Pattern` once `has_named` is known.
-                let mut named_fields = Vec::new();
-                let mut positional_fields = Vec::new();
-                let mut has_named = false;
-                let mut positional_index = 0usize;
-                for arg in args {
-                    match arg {
-                        ast::PatternArg::Named(field, pat) => {
-                            has_named = true;
-                            let mut field_place = place.clone();
-                            field_place.projection.push(bir::PlaceElem::Field(field.clone()));
-                            let lowered = self.lower_match_pattern(
-                                pat,
-                                &IncanType::Unknown,
-                                &field_place,
-                                arm_scope,
-                                arm,
-                                seen,
-                                saved_bindings,
-                            );
-                            named_fields.push((field.clone(), lowered));
-                        }
-                        ast::PatternArg::Positional(pat) => {
-                            let mut field_place = place.clone();
-                            field_place
-                                .projection
-                                .push(bir::PlaceElem::Field(positional_index.to_string()));
-                            positional_index += 1;
-                            let lowered = self.lower_match_pattern(
-                                pat,
-                                &IncanType::Unknown,
-                                &field_place,
-                                arm_scope,
-                                arm,
-                                seen,
-                                saved_bindings,
-                            );
-                            positional_fields.push(lowered);
-                        }
-                    }
-                }
-                if has_named {
-                    bir::Pattern::Struct {
-                        name: name.clone(),
-                        fields: named_fields,
-                    }
-                } else {
-                    bir::Pattern::Enum {
-                        name: String::new(),
-                        variant: name.clone(),
-                        fields: positional_fields,
-                    }
-                }
-            }
-            ast::Pattern::Group(inner) => {
-                self.lower_match_pattern(inner, expected_ty, place, arm_scope, arm, seen, saved_bindings)
-            }
-            ast::Pattern::Or(items) => {
-                let alternatives = items
-                    .iter()
-                    .map(|item| {
-                        self.lower_match_pattern(item, expected_ty, place, arm_scope, arm, seen, saved_bindings)
-                    })
-                    .collect();
-                bir::Pattern::Or(alternatives)
-            }
-        }
-    }
 }
 
 /// Return the first explicitly unsupported default statement, preserving the source span a direct consumer must
@@ -6222,6 +2463,30 @@ pub(crate) fn replacement_compatibility_body_ir_contribution()
         ],
     )
 }
+
+mod match_;
+
+mod closures;
+
+mod async_;
+
+mod literals;
+
+mod calls;
+
+mod operators;
+
+mod expr;
+
+mod defaults;
+
+mod assertions;
+
+mod comprehensions;
+
+mod control_flow;
+
+mod stmt;
 
 #[cfg(test)]
 mod tests;
